删除：[ gua_opencard120.js] 修改内容：[ cleancart_activity.js] 修改文件名：[ gua_opencard119.js -> gua_opencard124.js]
