+++ gua_opencard124.js
@@ -1,47 +1,63 @@
 /*
-雅诗兰黛奢雅邀请入会有礼 [gua_opencard120.js]
+3.30～4.08 约惠春天 好物来袭 [gua_opencard124.js]
 新增开卡脚本
 一次性脚本
 
-1.每邀请满3人50豆
-2.上限10组 1组3人（上限后退出脚本
-3.开1张卡
-4.已开卡的不算有效人数
+1.邀请一人20豆
+2.开10张卡 成功开1张 有机会获得10豆
+3.加购5京豆
+  (默认不加购 如需加购请设置环境变量[guaopencard_addSku124]为"true"
+4.抽奖 (默认不抽奖 如需抽奖请设置环境变量[guaopencard_draw124]为"3"
+填写要抽奖的次数 不足已自身次数为准
+guaopencard_draw124="3"
+填非数字会全都抽奖
 
 第一个账号助力作者 其他依次助力CK1
 第一个CK失效会退出脚本
 
 默认脚本不执行
 如需执行脚本请设置环境变量
-guaopencard120="true"
+guaopencard124="true"
 每个账号之间延迟 100=延迟100秒 0=延迟0秒会使用每3个账号延迟60秒
 guaopenwait_All 所有
-guaopenwait120="0"
+guaopenwait124="0"
 
 
 All变量适用
 ————————————————
-入口：[ 雅诗兰黛奢雅邀请入会有礼 (https://lzkjdz-isv.isvjcloud.com/esteelauder/inviteNew/activityPage?activityId=2203100037674501&inviterUuid=a336b1b2525f412ba91a27caf5998930)]
+入口：[ 3.30～4.08 约惠春天 好物来袭 (https://lzdz1-isv.isvjcloud.com/dingzhi/customized/common/activity?activityId=dzlhkkb306433997033e689d45c00a&shareUuid=25645ba5cde847d8bb3843699fc6d235)]
 
 请求太频繁会被黑ip
 过10分钟再执行
 
-cron:30 1 21-31/3 3 *
+cron:30 1 30,1-8/3 3 *
 ============Quantumultx===============
 [task_local]
-#雅诗兰黛奢雅邀请入会有礼
-30 1 21-31/3 3 * https://raw.githubusercontent.com/11111120/scripts/master/gua_opencard120.js, tag=雅诗兰黛奢雅邀请入会有礼, enabled=true
+#3.30～4.08 约惠春天 好物来袭
+30 1 30,1-8/3 3 * https://raw.githubusercontent.com/11111120/scripts/master/gua_opencard124.js, tag=3.30～4.08 约惠春天 好物来袭, enabled=true
 
 */
+let guaopencard_addSku = "false"
 let guaopencard = "false"
 let guaopenwait = "0"
+let guaopencard_draw = "0"
 
-const $ = new Env('雅诗兰黛奢雅邀请入会有礼');
+const $ = new Env('3.30～4.08 约惠春天 好物来袭');
 const jdCookieNode = $.isNode() ? require('./jdCookie.js') : '';
 const notify = $.isNode() ? require('./sendNotify') : '';
-//IOS等用户直接用NobyDa的jd cookie
 CryptoScripts()
 $.CryptoJS = $.isNode() ? require('crypto-js') : CryptoJS;
+let cleanCart = ''
+if($.isNode()){
+  try{
+    const fs = require('fs');
+    if (fs.existsSync('./cleancart_activity.js')) {
+      cleanCart = require('./cleancart_activity');
+    }
+  }catch(e){
+  }
+}
+//IOS等用户直接用NobyDa的jd cookie
 
 let cookiesArr = [],
     cookie = '';
@@ -54,15 +70,18 @@ if ($.isNode()) {
   cookiesArr = [$.getdata('CookieJD'), $.getdata('CookieJD2'), ...jsonParse($.getdata('CookiesJD') || "[]").map(item => item.cookie)].filter(item => !!item);
 }
 
-guaopencard = $.isNode() ? (process.env.guaopencard120 ? process.env.guaopencard120 : `${guaopencard}`) : ($.getdata('guaopencard120') ? $.getdata('guaopencard120') : `${guaopencard}`);
+guaopencard_addSku = $.isNode() ? (process.env.guaopencard_addSku124 ? process.env.guaopencard_addSku124 : `${guaopencard_addSku}`) : ($.getdata('guaopencard_addSku124') ? $.getdata('guaopencard_addSku124') : `${guaopencard_addSku}`);
+guaopencard_addSku = $.isNode() ? (process.env.guaopencard_addSku_All ? process.env.guaopencard_addSku_All : `${guaopencard_addSku}`) : ($.getdata('guaopencard_addSku_All') ? $.getdata('guaopencard_addSku_All') : `${guaopencard_addSku}`);
+guaopencard = $.isNode() ? (process.env.guaopencard124 ? process.env.guaopencard124 : `${guaopencard}`) : ($.getdata('guaopencard124') ? $.getdata('guaopencard124') : `${guaopencard}`);
 guaopencard = $.isNode() ? (process.env.guaopencard_All ? process.env.guaopencard_All : `${guaopencard}`) : ($.getdata('guaopencard_All') ? $.getdata('guaopencard_All') : `${guaopencard}`);
-guaopenwait = $.isNode() ? (process.env.guaopenwait120 ? process.env.guaopenwait120 : `${guaopenwait}`) : ($.getdata('guaopenwait120') ? $.getdata('guaopenwait120') : `${guaopenwait}`);
+guaopenwait = $.isNode() ? (process.env.guaopenwait124 ? process.env.guaopenwait124 : `${guaopenwait}`) : ($.getdata('guaopenwait124') ? $.getdata('guaopenwait124') : `${guaopenwait}`);
 guaopenwait = $.isNode() ? (process.env.guaopenwait_All ? process.env.guaopenwait_All : `${guaopenwait}`) : ($.getdata('guaopenwait_All') ? $.getdata('guaopenwait_All') : `${guaopenwait}`);
 guaopenwait = parseInt(guaopenwait, 10) || 0
+guaopencard_draw = $.isNode() ? (process.env.guaopencard_draw124 ? process.env.guaopencard_draw124 : guaopencard_draw) : ($.getdata('guaopencard_draw124') ? $.getdata('guaopencard_draw124') : guaopencard_draw);
+guaopencard_draw = $.isNode() ? (process.env.guaopencard_draw ? process.env.guaopencard_draw : guaopencard_draw) : ($.getdata('guaopencard_draw') ? $.getdata('guaopencard_draw') : guaopencard_draw);
 allMessage = ""
 message = ""
 $.hotFlag = false
-$.hasEnd = false
 $.outFlag = false
 $.activityEnd = false
 let lz_jdpin_token_cookie =''
@@ -70,7 +89,7 @@ let activityCookie =''
 !(async () => {
   if ($.isNode()) {
     if(guaopencard+"" != "true"){
-      console.log('如需执行脚本请设置环境变量[guaopencard120]为"true"')
+      console.log('如需执行脚本请设置环境变量[guaopencard124]为"true"')
     }
     if(guaopencard+"" != "true"){
       return
@@ -82,16 +101,15 @@ let activityCookie =''
     });
     return;
   }
-  $.joinStatus = false
-  $.activityId = "2203100037674501"
-  $.shareUuid = "46020214e6174cb3a748babecada2a7a"
-  console.log(`入口:\nhttps://lzkjdz-isv.isvjcloud.com/esteelauder/inviteNew/activityPage?activityId=2203100037674501&inviterUuid=${$.shareUuid}`)
-  let shareUuidArr = ["46020214e6174cb3a748babecada2a7a","31dce333f9f547d19b611424b3e12dec","042557422be24d25959c8d2963c73cbe","5cff9128794641ce90fa4f48b134eaff","5bed53f7c1fa4059b7c542cb6f7f734a","1c2530ff29fc494f898e42795d49c4c7","65ac61c216df45899b4397800c8a5cb1","890217a3c0f947ad9432916619c8d8e0","c6e51f0cc85f4d58a23323c0204cf631","77e491ac12774ea7bdac4ecfb6ed1237"]
+  $.activityId = "dzlhkkb306433997033e689d45c00a"
+  $.shareUuid = "25645ba5cde847d8bb3843699fc6d235"
+  console.log(`入口:\nhttps://lzdz1-isv.isvjcloud.com/dingzhi/customized/common/activity?activityId=${$.activityId}&shareUuid=${$.shareUuid}`)
+  let shareUuidArr = [$.shareUuid,"148987f9e41443beb9f7a18d4dffde54","1a865bacc23042808795911221ec5698","c784aac0d5984bc7bcba349dd9e5f8e1","61e316a03d7242b588b9c70971bc715f","b76f0e679ead4d9ba741baa17f70b063","a4a31786631448dd8bff830c90ae9833","2184dc8f09e14113bdf136d6fdaa6900","377a7a5644834651939ad07b64dd6237","d3bbf1cb7ad5480081f4a11660802e30"]
   let s = Math.floor((Math.random()*10))
   let n = 0
-  n = Math.floor((Math.random()*shareUuidArr.length))
+  if(s >= 1 && s<= 6) n = Math.floor((Math.random()*shareUuidArr.length))
   $.shareUuid = shareUuidArr[n] ? shareUuidArr[n] : $.shareUuid
-  
+
   for (let i = 0; i < cookiesArr.length; i++) {
     cookie = cookiesArr[i];
     if (cookie) {
@@ -99,28 +117,15 @@ let activityCookie =''
       $.index = i + 1;
       message = ""
       $.bean = 0
+      $.hotFlag = false
       $.nickName = '';
       console.log(`\n\n******开始【京东账号${$.index}】${$.nickName || $.UserName}*********\n`);
       await getUA()
       await run();
       if(i == 0 && !$.actorUuid) break
-      if($.outFlag || $.activityEnd || $.hasEnd) break
-      // break
+      if($.outFlag || $.activityEnd) break
     }
   }
-  cookie = cookiesArr[0];
-  if (cookie && $.joinStatus && !$.outFlag && !$.activityEnd) {
-    $.UserName = decodeURIComponent(cookie.match(/pt_pin=([^; ]+)(?=;?)/) && cookie.match(/pt_pin=([^; ]+)(?=;?)/)[1])
-    $.index = 1;
-    message = ""
-    $.bean = 0
-    $.nickName = '';
-    console.log(`\n\n******开始【京东账号${$.index}】${$.nickName || $.UserName}*********\n`);
-    await $.wait(parseInt(Math.random() * 2000 + 4000, 10))
-    await getUA()
-    await run();
-  }
-  
   if($.outFlag) {
     let msg = '此ip已被限制，请过10分钟后再执行脚本'
     $.msg($.name, ``, `${msg}`);
@@ -128,7 +133,7 @@ let activityCookie =''
   }
   if(allMessage){
     $.msg($.name, ``, `${allMessage}`);
-    if ($.isNode()) await notify.sendNotify(`${$.name}`, `${allMessage}`);
+    // if ($.isNode()) await notify.sendNotify(`${$.name}`, `${allMessage}`);
   }
 })()
     .catch((e) => $.logErr(e))
@@ -137,14 +142,14 @@ let activityCookie =''
 
 async function run() {
   try {
+    $.hasEnd = true
     $.endTime = 0
-    $.assistCount = 0
     lz_jdpin_token_cookie = ''
     $.Token = ''
     $.Pin = ''
     let flag = false
     await takePostRequest('isvObfuscator');
-    if ($.Token == '') {
+    if($.Token == ''){
       console.log('获取[token]失败！')
       return
     }
@@ -152,79 +157,150 @@ async function run() {
     if (activityCookie == '') {
       console.log(`获取cookie失败`); return;
     }
-    if ($.activityEnd === true) {
+    if($.activityEnd === true){
       console.log('活动结束')
       return
     }
-    if ($.outFlag) {
+    if($.outFlag){
       console.log('此ip已被限制，请过10分钟后再执行脚本\n')
       return
     }
+    await takePostRequest('getSimpleActInfoVo');
     await takePostRequest('getMyPing');
-    if (!$.Pin) {
+    if(!$.Pin){
       console.log('获取[Pin]失败！')
       return
     }
     await takePostRequest('accessLogWithAD');
+    await takePostRequest('getUserInfo');
     await takePostRequest('activityContent');
-    // if($.hasEnd === true || Date.now() > $.endTime || Date.now()< $.startTimes){
-    //   $.activityEnd = true
-    //   console.log('不在活动范围内')
-    //   return
-    // }
-    if (!$.actorUuid) {
+    if($.hotFlag) return
+    if(!$.actorUuid){
       console.log('获取不到[actorUuid]退出执行，请重新执行')
       return
     }
-    if ($.openStatus == 0) {
-      console.log('开卡')
-      $.joinVenderId = 1000376745
-      for (let i = 0; i < Array(5).length; i++) {
-        if (i > 0) console.log(`第${i}次 重新开卡`)
-        await joinShop()
-        if ($.errorJoinShop.indexOf('活动太火爆，请稍后再试') == -1) {
-          break
+    if($.hasEnd === true || Date.now() > $.endTime){
+      $.activityEnd = true
+      console.log('活动结束')
+      return
+    }
+    await takePostRequest('drawContent');
+    await $.wait(1000)
+    $.openList = []
+    $.allOpenCard = false
+    await takePostRequest('info');
+    await takePostRequest('checkOpenCard');
+    console.log($.actorUuid)
+    // return
+    if($.allOpenCard == false){
+      console.log('开卡任务')
+      for(o of $.openList){
+        $.openCard = false
+        if(o.status == 0){
+          flag = true
+          $.shopactivityId = ''
+          $.joinVenderId = o.venderId
+          await getshopactivityId()
+          for (let i = 0; i < Array(5).length; i++) {
+            if (i > 0) console.log(`第${i}次 重新开卡`)
+            await joinShop()
+            if ($.errorJoinShop.indexOf('活动太火爆，请稍后再试') == -1) {
+              break
+            }
+          }
+          if ($.errorJoinShop.indexOf('活动太火爆，请稍后再试') > -1) {
+            console.log("开卡失败❌ ，重新执行脚本")
+            allMessage += `【账号${$.index}】开卡失败❌ ，重新执行脚本\n`
+          } else {
+            $.joinStatus = true
+          }
+          await takePostRequest('activityContent');
+          await takePostRequest('drawContent');
+          await takePostRequest('checkOpenCard');
+          await $.wait(parseInt(Math.random() * 2000 + 3000, 10))
         }
       }
-      if ($.errorJoinShop.indexOf('活动太火爆，请稍后再试') > -1) {
-        console.log("开卡失败❌ ，重新执行脚本")
-        allMessage += `【账号${$.index}】开卡失败❌ ，重新执行脚本\n`
-      } else {
-        $.joinStatus = true
+    }else{
+      console.log('已全部开卡')
+    }
+    
+    $.log("关注: " + $.followShop)
+    if(!$.followShop && !$.outFlag){
+      flag = true
+      await takePostRequest('followShop');
+      await $.wait(parseInt(Math.random() * 2000 + 3000, 10))
+    }
+
+    $.yaoqing = false
+    await takePostRequest('邀请');
+    if($.yaoqing){
+      await takePostRequest('助力');
+    }
+    $.log("加购: " + $.addCart)
+    if(!$.addCart && !$.outFlag){
+      if(guaopencard_addSku+"" == "true"){
+        flag = true
+        let goodsArr = []
+        if(cleanCart){
+          goodsArr = await cleanCart.clean(cookie,'https://jd.11111118/jdcleancatr_21102717','')
+          if(goodsArr !== false) await $.wait(parseInt(Math.random() * 1000 + 4000, 10))
+        }
+        await takePostRequest('addCart');
+        await $.wait(parseInt(Math.random() * 2000 + 4000, 10))
+        if(cleanCart && goodsArr !== false){
+          // await $.wait(parseInt(Math.random() * 1000 + 4000, 10))
+          await cleanCart.clean(cookie,'https://jd.11111118/jdcleancatr_21102717',goodsArr || [ ])
+        }
+      }else{
+        console.log('如需加购请设置环境变量[guaopencard_addSku124]为"true"');
       }
+    }
+    if(flag){
       await takePostRequest('activityContent');
-      await $.wait(parseInt(Math.random() * 2000 + 2000, 10))
     }
-    console.log($.openStatus === 1 ? "已开卡" : $.openStatus === 2 ? "已开卡" : $.openStatus === 0 ? "未开卡" : "未知-" + $.openStatus)
-    console.log($.helpStatus === 1 ? "助力成功" : $.helpStatus === 0 ? "助力失败" : "助力失败-" + $.helpStatus)
-    if ($.index == 1) {
-      $.helpCount = $.assistCount
-    }else if($.helpStatus == 1){
-      $.helpCount++;
-    }
-    if ($.drawTimes > 0) {
-      let count = parseInt($.drawTimes / 1)
-      for (m = 1; count--; m++) {
-        await takePostRequest('convertPrize');
+    console.log(`${$.score}值`)
+    if(guaopencard_draw+"" !== "0"){
+      $.runFalag = true
+      let count = parseInt($.score/100)
+      guaopencard_draw = parseInt(guaopencard_draw, 10)
+      if(count > guaopencard_draw) count = guaopencard_draw
+      console.log(`抽奖次数为:${count}`)
+      for(m=1;count--;m++){
+        console.log(`第${m}次抽奖`)
+        await takePostRequest('抽奖');
+        if($.runFalag == false) break
+        if(Number(count) <= 0) break
+        if(m >= 10){
+          console.log("抽奖太多次，多余的次数请再执行脚本")
+          break
+        }
+        await $.wait(parseInt(Math.random() * 2000 + 2000, 10))
       }
+    }else console.log('如需抽奖请设置环境变量[guaopencard_draw124]为"3" 3为次数');
+    
+    await $.wait(parseInt(Math.random() * 1000 + 2000, 10))
+    await takePostRequest('getDrawRecordHasCoupon');
+    await takePostRequest('getShareRecord');
+    if($.outFlag){
+      console.log('此ip已被限制，请过10分钟后再执行脚本\n')
+      return
     }
-    console.log(`【账号${$.index}】助力人数：${$.assistCount}${$.index != 1 && " 【账号1】助力人数：" + $.helpCount || ""}`)
-    if ($.helpCount >= 10 * 3) $.hasEnd = true
     console.log($.actorUuid)
     console.log(`当前助力:${$.shareUuid}`)
-    if ($.index == 1) {
+    if($.index == 1){
       $.shareUuid = $.actorUuid
       console.log(`后面的号都会助力:${$.shareUuid}`)
     }
-    await $.wait(parseInt(Math.random() * 2000 + 5000, 10))
-    if (guaopenwait) {
-      if ($.index != cookiesArr.length) {
+    await $.wait(parseInt(Math.random() * 1000 + 5000, 10))
+    if(flag) await $.wait(parseInt(Math.random() * 1000 + 10000, 10))
+    if(guaopenwait){
+      if($.index != cookiesArr.length){
         console.log(`等待${guaopenwait}秒`)
         await $.wait(parseInt(guaopenwait, 10) * 1000)
       }
-    } else {
-      if ($.index % 3 == 0) console.log('休息1分钟，别被黑ip了\n可持续发展')
-      if ($.index % 3 == 0) await $.wait(parseInt(Math.random() * 5000 + 60000, 10))
+    }else{
+      if($.index % 3 == 0) console.log('休息1分钟，别被黑ip了\n可持续发展')
+      if($.index % 3 == 0) await $.wait(parseInt(Math.random() * 5000 + 60000, 10))
     }
   } catch (e) {
     console.log(e)
@@ -232,68 +308,139 @@ async function run() {
 }
 
 async function takePostRequest(type) {
-  if ($.outFlag) return
-  let domain = 'https://lzkjdz-isv.isvjcloud.com';
+  if($.outFlag) return
+  let domain = 'https://lzdz1-isv.isvjcloud.com';
   let body = ``;
   let method = 'POST'
   let admJson = ''
   switch (type) {
     case 'isvObfuscator':
       url = `https://api.m.jd.com/client.action?functionId=isvObfuscator`;
-      body = `body=%7B%22url%22%3A%22https%3A//lzkjdz-isv.isvjcloud.com%22%2C%22id%22%3A%22%22%7D&uuid=41bfd2e1dc84ad874fc29dc7c7e75f1a5e8543b7&client=apple&clientVersion=10.1.4&st=1647806351423&sv=102&sign=bc8986c4677c3d1c1185068553e15ed8`;
+      body = `body=%7B%22url%22%3A%22https%3A//lzdz1-isv.isvjcloud.com%22%2C%22id%22%3A%22%22%7D&uuid=9bedc1528ce297b982ce8bb0a8785ae48c18500e&client=apple&clientVersion=10.1.4&st=1646363852923&sv=111&sign=24a34053625f61226d89fd9402c42f6a`;
       break;
-    case 'getMyPing':
-      url = `${domain}/customer/getMyPing`;
-      body = `token=${$.Token}&fromType=APP&userId=1000376745&pin=`;
-      break;
-    case 'getSimpleActInfoVo':
-      url = `${domain}/common/brand/getSimpleActInfoVo`;
-      body = `activityId=${$.activityId}`;
-      break;
-    case 'accessLogWithAD':
-      url = `${domain}/common/accessLogWithAD`;
-      let pageurl = `https://lzkjdz-isv.isvjcloud.com/esteelauder/inviteNew/activityPage?activityId=2203100037674501&inviterUuid=${$.shareUuid}`
-      body = `venderId=1000376745&code=99&pin=${encodeURIComponent($.Pin)}&activityId=${$.activityId}&pageUrl=${encodeURIComponent(pageurl)}`
-      break;
-    case 'activityContent':
-      url = `${domain}/esteelauder/inviteNew/activityContent`;
-      body = `activityId=${$.activityId}&pin=${encodeURIComponent($.Pin)}&inviterUuid=${$.shareUuid}`
-      break;
-    case 'convertPrize':
-      url = `${domain}/esteelauder/inviteNew/convertPrize`;
-      body = `activityId=${$.activityId}&pin=${encodeURIComponent($.Pin)}&uuid=${$.shareUuid}`
-      break;
-    default:
-      console.log(`错误${type}`);
-  }
-  let myRequest = getPostRequest(url, body, method);
-  // console.log(myRequest)
-  return new Promise(async resolve => {
-    $.post(myRequest, (err, resp, data) => {
-      try {
-        setActivityCookie(resp)
-        if (err) {
-          if (resp && typeof resp.statusCode != 'undefined') {
-            if (resp.statusCode == 493) {
-              console.log('此ip已被限制，请过10分钟后再执行脚本\n')
-              $.outFlag = true
-            }
-          }
-          console.log(`${$.toStr(err, err)}`)
-          console.log(`${type} API请求失败，请检查网路重试`)
-        } else {
-          dealReturn(type, data);
+      case 'getSimpleActInfoVo':
+        url = `${domain}/dz/common/getSimpleActInfoVo`;
+        body = `activityId=${$.activityId}`;
+        break;
+      case 'getMyPing':
+        url = `${domain}/customer/getMyPing`;
+        body = `userId=${$.shopId || $.venderId || ''}&token=${$.Token}&fromType=APP`;
+        break;
+      case 'accessLogWithAD':
+        url = `${domain}/common/accessLogWithAD`;
+        let pageurl = `${domain}/drawCenter/activity?activityId=${$.activityId}&shareUuid=${$.shareUuid}`
+        body = `venderId=${$.shopId || $.venderId || ''}&code=99&pin=${encodeURIComponent($.Pin)}&activityId=${$.activityId}&pageUrl=${encodeURIComponent(pageurl)}&subType=app&adSource=`
+        break;
+      case 'getUserInfo':
+        url = `${domain}/wxActionCommon/getUserInfo`;
+        body = `pin=${encodeURIComponent($.Pin)}`;
+        break;
+      case 'activityContent':
+        url = `${domain}/dingzhi/linkgame/activity/content`;
+        body = `activityId=${$.activityId}&pin=${encodeURIComponent($.Pin)}&pinImg=${encodeURIComponent($.attrTouXiang)}&nick=${encodeURIComponent($.nickname)}&cjyxPin=&cjhyPin=&shareUuid=${$.shareUuid}`
+        break;
+      case 'drawContent':
+        url = `${domain}/dingzhi/taskact/common/drawContent`;
+        body = `activityId=${$.activityId}&pin=${encodeURIComponent($.Pin)}`
+        break;
+      case 'checkOpenCard':
+        url = `${domain}/dingzhi/linkgame/checkOpenCard`;
+        body = `activityId=${$.activityId}&pin=${encodeURIComponent($.Pin)}&shareUuid=${$.shareUuid}`
+        break;
+      case 'info':
+        url = `${domain}/dingzhi/linkgame/task/opencard/info`;
+        body = `activityId=${$.activityId}&pin=${encodeURIComponent($.Pin)}&actorUuid=${$.actorUuid}`
+        break;
+      case 'startDraw':
+        url = `${domain}/joint/order/draw`;
+        body = `activityId=${$.activityId}&pin=${encodeURIComponent($.Pin)}&actorUuid=${$.actorUuid}&drawType=1`
+        break;
+      case 'followShop':
+        url = `${domain}/dingzhi/opencard/follow/shop`;
+        // url = `${domain}/dingzhi/dz/openCard/saveTask`;
+        body = `activityId=${$.activityId}&pin=${encodeURIComponent($.Pin)}`
+        break;
+      case 'sign':
+      case 'addCart':
+      case 'browseGoods':
+        url = `${domain}/dingzhi/opencard/${type}`;
+        body = `activityId=${$.activityId}&pin=${encodeURIComponent($.Pin)}`
+        if(type == 'browseGoods') body += `&value=${$.visitSkuValue}`
+        break;
+      case '邀请':
+      case '助力':
+        if(type == '助力'){
+          url = `${domain}/dingzhi/linkgame/assist`;
+        }else{
+          url = `${domain}/dingzhi/linkgame/assist/status`;
         }
-      } catch (e) {
-        // console.log(data);
-        console.log(e, resp)
-      } finally {
-        resolve();
-      }
+        body = `activityId=${$.activityId}&pin=${encodeURIComponent($.Pin)}&shareUuid=${$.shareUuid}`
+        break;
+      case 'viewVideo':
+      case 'visitSku':
+      case 'toShop':
+      case 'addSku':
+        url = `${domain}/dingzhi/opencard/${type}`;
+        let taskType = ''
+        let taskValue = ''
+        if(type == 'viewVideo'){
+          taskType = 31
+          taskValue = 31
+        }else if(type == 'visitSku'){
+          taskType = 5
+          taskValue = $.visitSkuValue || 5
+        }else if(type == 'toShop'){
+          taskType = 14
+          taskValue = $.toShopValue || 14
+        }else if(type == 'addSku'){
+          taskType = 2
+          taskValue = $.addSkuValue || 2
+        }
+        body = `activityId=${$.activityId}&pin=${encodeURIComponent($.Pin)}&actorUuid=${$.actorUuid}&taskType=${taskType}&taskValue=${taskValue}`
+        break;
+      case 'getDrawRecordHasCoupon':
+        url = `${domain}/dingzhi/linkgame/draw/record`;
+        body = `activityId=${$.activityId}&pin=${encodeURIComponent($.Pin)}&actorUuid=${$.actorUuid}`
+        break;
+      case 'getShareRecord':
+        url = `${domain}/dingzhi/linkgame/help/list`;
+        body = `activityId=${$.activityId}&pin=${encodeURIComponent($.Pin)}`
+        break;
+      case '抽奖':
+        url = `${domain}/dingzhi/opencard/draw`;
+        body = `activityId=${$.activityId}&actorUuid=${$.actorUuid}&pin=${encodeURIComponent($.Pin)}`
+        break;
+      default:
+        console.log(`错误${type}`);
+    }
+    let myRequest = getPostRequest(url, body, method);
+    // console.log(myRequest)
+    return new Promise(async resolve => {
+      $.post(myRequest, (err, resp, data) => {
+        try {
+          setActivityCookie(resp)
+          if (err) {
+            if(resp && typeof resp.statusCode != 'undefined'){
+              if(resp.statusCode == 493){
+                console.log('此ip已被限制，请过10分钟后再执行脚本\n')
+                $.outFlag = true
+              }
+            }
+            console.log(`${$.toStr(err,err)}`)
+            console.log(`${type} API请求失败，请检查网路重试`)
+          } else {
+            dealReturn(type, data);
+          }
+        } catch (e) {
+          // console.log(data);
+          console.log(e, resp)
+        } finally {
+          resolve();
+        }
+      })
     })
-  })
-}
-
+  }
+  
 async function dealReturn(type, data) {
   let res = ''
   try {
@@ -322,6 +469,20 @@ async function dealReturn(type, data) {
           console.log(data)
         }
         break;
+      case 'getSimpleActInfoVo':
+        if(typeof res == 'object'){
+          if(res.result && res.result === true){
+            if(typeof res.data.shopId != 'undefined') $.shopId = res.data.shopId
+            if(typeof res.data.venderId != 'undefined') $.venderId = res.data.venderId
+          }else if(res.errorMessage){
+            console.log(`${type} ${res.errorMessage || ''}`)
+          }else{
+            console.log(`${type} ${data}`)
+          }
+        }else{
+          console.log(`${type} ${data}`)
+        }
+        break;
       case 'getMyPing':
         if(typeof res == 'object'){
           if(res.result && res.result === true){
@@ -336,20 +497,11 @@ async function dealReturn(type, data) {
           console.log(`${type} ${data}`)
         }
         break;
-      case 'activityContent':
+      case 'getUserInfo':
         if(typeof res == 'object'){
           if(res.result && res.result === true){
-            $.actorUuid = res.data.userRecord.uuid || ''
-            $.helpStatus = res.data.userRecord.assistStatus || 0
-            $.openStatus = res.data.userRecord.openCardStatus || 0
-            $.assistCount = res.data.userRecord.assistCount || 0
-            $.drawTimes = res.data.drawTimes || 0
-            if(res.data.sendBeanNum){
-              console.log(`获得${res.data.sendBeanNum}豆`)
-              allMessage += `【账号${$.index}】获得${res.data.sendBeanNum}豆\n`
-            }
+            if(res.data && typeof res.data.yunMidImageUrl != 'undefined') $.attrTouXiang = res.data.yunMidImageUrl || "https://img10.360buyimg.com/imgzone/jfs/t1/7020/27/13511/6142/5c5138d8E4df2e764/5a1216a3a5043c5d.png"
           }else if(res.errorMessage){
-            if(res.errorMessage.indexOf("结束") > -1) $.activityEnd = true
             console.log(`${type} ${res.errorMessage || ''}`)
           }else{
             console.log(`${type} ${data}`)
@@ -358,15 +510,18 @@ async function dealReturn(type, data) {
           console.log(`${type} ${data}`)
         }
         break;
-      case 'convertPrize':
+      case 'activityContent':
         if(typeof res == 'object'){
           if(res.result && res.result === true){
-            console.log(res.data.giftName)
-            if(res.data.giftName.indexOf('京豆') == -1 && res.data.giftName.indexOf('谢谢') == -1){
-              allMessage += `【账号${$.index}】获得${res.data.giftName}\n`
-            }
-          }else if(res.errorMessage || res.msg){
-            console.log(`${type} ${res.errorMessage || res.msg || ''}`)
+            $.endTime = res.data.endTime || (res.data.activityVo && res.data.activityVo.endTime) || res.data.activity.endTime || 0
+            $.hasEnd = res.data.isEnd || false
+            $.drawCount = res.data.actor.drawCount || 0
+            $.point = res.data.actor.point || 0
+            $.score = res.data.actor.score || 0
+            $.actorUuid = res.data.actor.actorUuid || ''
+            $.followShop = res.data.actor.followShopStatus || ''
+          }else if(res.errorMessage){
+            console.log(`${type} ${res.errorMessage || ''}`)
           }else{
             console.log(`${type} ${data}`)
           }
@@ -374,6 +529,164 @@ async function dealReturn(type, data) {
           console.log(`${type} ${data}`)
         }
         break;
+      case 'info':
+        if(typeof res == 'object'){
+          if(res.result && res.result === true){
+            // $.drawCount = res.data.drawCount || 0
+            $.addCart = res.data.addCart || false
+            // $.followShop = res.data.followShop || false
+            // $.sign = res.data.isSignStatus || false
+            // $.visitSku = res.data.visitSku || false
+            // $.visitSkuList = res.data.visitSkuList || []
+          }else if(res.errorMessage){
+            console.log(`${type} ${res.errorMessage || ''}`)
+          }else{
+            console.log(`${type} ${data}`)
+          }
+        }else{
+          console.log(`${type} ${data}`)
+        }
+        break;
+      case 'checkOpenCard':
+        if(typeof res == 'object'){
+          if(res.result && res.result === true){
+            let cardList1 = res.data.cardList1 || []
+            let cardList2 = res.data.cardList2 || []
+            let cardList = res.data.cardList || []
+            let openCardList = res.data.openCardList || []
+            $.openList = [...cardList,...cardList1,...cardList2,...openCardList]
+            $.allOpenCard = res.data.allOpenCard || res.data.isOpenCardStatus || false
+            $.openCardScore1 = res.data.score1 || 0
+            $.openCardScore2 = res.data.score2 || 0
+            $.drawScore = res.data.drawScore || 0
+            if(res.data.beans || res.data.addBeanNum) console.log(`开卡获得:${res.data.beans || res.data.addBeanNum}豆`)
+          }else if(res.errorMessage){
+            console.log(`${type} ${res.errorMessage || ''}`)
+          }else{
+            console.log(`${type} ${data}`)
+          }
+        }else{
+          console.log(`${type} ${data}`)
+        }
+        break;
+      case 'startDraw':
+      case 'followShop':
+      case 'viewVideo':
+      case 'visitSku':
+      case 'toShop':
+      case 'addSku':
+      case 'sign':
+      case 'addCart':
+      case 'browseGoods':
+      case '抽奖':
+        if(typeof res == 'object'){
+          if(res.result && res.result === true){
+            if(typeof res.data == 'object'){
+              let msg = ''
+              let title = '抽奖'
+              if(res.data.addBeanNum){
+                msg = `${res.data.addBeanNum}京豆`
+              }
+              if(res.data.addPoint){
+                msg += ` ${res.data.addPoint}游戏机会`
+              }
+              if(type == 'followShop'){
+                title = '关注'
+                if(res.data.beanNumMember && res.data.assistSendStatus){
+                  msg += ` 额外获得:${res.data.beanNumMember}京豆`
+                }
+              }else if(type == 'addSku' || type == 'addCart'){
+                title = '加购'
+              }else if(type == 'viewVideo'){
+                title = '热门文章'
+              }else if(type == 'toShop'){
+                title = '浏览店铺'
+              }else if(type == 'visitSku' || type == 'browseGoods'){
+                title = '浏览商品'
+              }else if(type == 'sign'){
+                title = '签到'
+              }else{
+                msg = res.data.drawOk == true && (res.data.drawInfoType == 6 && res.data.name || '') || '空气💨'
+              }
+              if(!msg){
+                msg = '空气💨'
+              }
+              console.log(`${title}获得:${msg || data}`)
+            }else{
+              console.log(`${type} ${data}`)
+            }
+          }else if(res.errorMessage){
+            $.runFalag = false;
+            console.log(`${type} ${res.errorMessage || ''}`)
+          }else{
+            console.log(`${type} ${data}`)
+          }
+        }else{
+          console.log(`${type} ${data}`)
+        }
+        break;
+      case 'getDrawRecordHasCoupon':
+        if(typeof res == 'object'){
+          if(res.result && res.result === true){
+            console.log(`我的奖品：`)
+            let num = 0
+            let value = 0
+            for(let i in res.data.recordList){
+              let item = res.data.recordList[i]
+              if(item.infoName == '20京豆' && item.drawStatus == 0){
+                num++
+                value = item.infoName.replace('京豆','')
+              }else{
+                console.log(`${item.infoType != 10 && item.value && item.value +':' || ''}${item.infoName}`)
+              }
+            }
+            if(num > 0) console.log(`邀请好友(${num}):${num*parseInt(value, 10) || 30}京豆`)
+          }else if(res.errorMessage){
+            console.log(`${type} ${res.errorMessage || ''}`)
+          }else{
+            console.log(`${type} ${data}`)
+          }
+        }else{
+          console.log(`${type} ${data}`)
+        }
+        break;
+      case 'getShareRecord':
+        if(typeof res == 'object'){
+          if(res.result && res.result === true && res.data){
+            $.ShareCount = res.data.shareList.length
+            $.log(`=========== 你邀请了:${$.ShareCount}个\n由于接口数据只有30个 故邀请大于30个的需要自行判断\n`)
+          }else if(res.errorMessage){
+            console.log(`${type} ${res.errorMessage || ''}`)
+          }else{
+            console.log(`${type} ${data}`)
+          }
+        }else{
+          console.log(`${type} ${data}`)
+        }
+        break;
+      case '邀请':
+      case '助力':
+        // console.log(data)
+        if(typeof res == 'object'){
+          if(res.data.status == 200){
+            if(type == '助力'){
+              console.log('助力成功')
+            }else{
+              $.yaoqing = true
+            }
+          }else if(res.data.status == 105){
+            console.log('已经助力过')
+          }else if(res.data.status == 104){
+            console.log('已经助力其他人')
+          }else if(res.data.status == 101){
+            // console.log('已经助力过')
+          }else{
+            console.log(data)
+          }
+        }else{
+          console.log(`${type} ${data}`)
+        }
+
       case 'accessLogWithAD':
       case 'drawContent':
         break;
@@ -403,8 +716,8 @@ function getPostRequest(url, body, method="POST") {
     "User-Agent": $.UA,
     "X-Requested-With": "XMLHttpRequest"
   }
-  if(url.indexOf('https://lzkjdz-isv.isvjcloud.com') > -1){
-    headers["Referer"] = `https://lzkjdz-isv.isvjcloud.com/esteelauder/inviteNew/activityPage?activityId=2203100037674501&inviterUuid=${$.shareUuid}`
+  if(url.indexOf('https://lzdz1-isv.isvjcloud.com') > -1){
+    headers["Referer"] = `https://lzdz1-isv.isvjcloud.com/dingzhi/customized/common/activity?activityId=${$.activityId}&shareUuid=${$.shareUuid}`
     headers["Cookie"] = `${lz_jdpin_token_cookie && lz_jdpin_token_cookie || ''}${$.Pin && "AUTH_C_USER=" + $.Pin + ";" || ""}${activityCookie}`
   }
   // console.log(headers)
@@ -415,15 +728,9 @@ function getPostRequest(url, body, method="POST") {
 function getCk() {
   return new Promise(resolve => {
     let get = {
-      url:`https://lzkjdz-isv.isvjcloud.com/wxCommonInfo/token`,
+      url:`https://lzdz1-isv.isvjcloud.com/dingzhi/customized/common/activity?activityId=${$.activityId}&shareUuid=${$.shareUuid}`,
+      followRedirect:false,
       headers: {
-        "Accept": "application/json, text/plain, */*",
-        "Accept-Encoding": "gzip, deflate, br",
-        "Accept-Language": "zh-cn",
-        "Connection": "keep-alive",
-        "Content-Type": "application/x-www-form-urlencoded",
-        "Cookie": cookie,
-        "Referer": `https://lzkjdz-isv.isvjcloud.com/esteelauder/inviteNew/activityPage?activityId=2203100037674501&inviterUuid=${$.shareUuid}`,
         "User-Agent": $.UA,
       },
       timeout:30000
@@ -490,13 +797,11 @@ function randomString(e) {
   return n
 }
 
+
 function joinShop() {
   if(!$.joinVenderId) return
   return new Promise(async resolve => {
-    $.shopactivityId = ''
     $.errorJoinShop = ''
-    await $.wait(1000)
-    await getshopactivityId()
     let activityId = ``
     if($.shopactivityId) activityId = `,"activityId":${$.shopactivityId}`
     let body = `{"venderId":"${$.joinVenderId}","shopId":"${$.joinVenderId}","bindByVerifyCodeFlag":1,"registerExtend":{},"writeChildFlag":0${activityId},"channel":401}`
@@ -590,14 +895,12 @@ function jsonParse(str) {
   }
 }
 
-var _0xodp='jsjiami.com.v6',_0xodp_=['‮_0xodp'],_0xb853=[_0xodp,'UsOxwosIOw==','GS/CvcO3IA==','Z0Q5w7Fn','eE4Nw4Z5','dMO3wrUFKA==','w7wUw5Jzwqk=','ZcOVwokZAQ==','GEozwpTCsw==','EQfChcOMLg==','Py3DkWjDk1M=','w5k+wpVobcOK','wqXCi8Krw43Cpw==','wrjDmmDDmcOZ','w7Qhw6g=','w7PCosOBR2I=','w4fDmMKgwoLCvA==','LwvCo8O9EEoxI2o=','w5LDmcKCwqjCiw==','w5kFw650wqlxDsKyb8KzcTTDicKR','J0/Cmgc=','wqMIb2fDrxo=','EsK3GcOUw5rCoMKzSsKhw5DCuMKRBQ==','wo5fbMKYMA==','NijCsMKbLTgzw7NXbg==','wptCw6XCl8Ouw7USfyV0wro1w6tq','w5nDmsKdwq/Cnw==','wq/CsUjCksOvXA==','OjLCt8KW','w6LCsVRvNMKbSMObw67Ci1MQwr/DiMKx','wp83EcOKQw==','NsKAwr4=','wrnDl8KvSsO4','w7zCrVlCEMK7SQ==','w5DDj8Oow6VN','w7HDmMKewo7CiA==','a1jDrsKRwrs=','w4bCoildw4I=','w7rDksOiw6pL','w7sKw4ZswrQ=','MFPDvsK3Xg==','w7oWw5tywoY=','wpTDucKBbsOa','w449JMOVSA==','BlLChw7DhkzCmAA=','w5bChyo8wqbCsQ==','OSPDrHjDjg==','wpXCvsKhw6vCosOpX8Og','w7nCugTDlnE=','wqgFAMKhw4g=','J2M8wqLCng==','WsOSwqQtJQ==','wpYaJ8KJw70=','w4jCr8OrVz4=','OcK5CMOdw4I=','w7t8XTA6','woHCoU3Cj8O7','w5Ivw7xUwoQ=','wozCklM=','wqbCp07CkA==','PWk0wrPCkA==','w6nCqUNyFsK7','PTzCrg==','w6EXwp9YSg==','FsKQb8ORbQ==','wqbDisKvwot3','JGkIwqPCvQ==','w4gFw4dFwo0=','IGTDnMK3XQ==','wpbCmMK+w5XCoA==','w7PDhcOzw5hC','wpbCkk9gMw==','w5h6WQg7w6E=','w4UUKsODciZ+','KWAwwoY=','w5nDhMO9XMOs','LDoX','wpbClElg','wqk+FcKYw6E=','MwrDqi4BZg==','YE3CoBfDug==','woRMw7TDlxpLw5Fr','BcK0I8OFw4bCvcKLSA==','IzTCt8OIEg==','NjozwrUbwowobw==','w7YpwoJ/bcONwrXCkg==','B0cNwrTCg25sw6I=','woQuDcKSw706YsKA','BSHDnm/Dr3YGGsKeSA==','w4HDlEHDq2bDpsONUA==','w4MeDcOCciEww4U=','w4Q5G8OPQCs=','wpbCkndxL8OfGyw=','OQrCqcO7F1c=','Zzo+YQXDlA==','w7jDgsOBScOe','JcKWVMOKTsOK','B0c1wqXCnw==','wqbCrUnCn8O+dw==','wqQvXkHDkA==','w5QeMMOVYTw=','YSrCkcKo','w5PDgMOSTcO1','wq/DiV7DhcOE','w6zCjhTCsMKhDC4cwqt6','wqwUw610wqAbwo8KEcO9KHEuLcOSwrY=','w77CpsO1UUXCksO1XcKHRlxIQj0+eMOa','IA3DqsObOA8lLjUrwqpuwoEXS8Ofw4hzEsOHwo8PwoU=','w58FKsOGc3Jxwo1qwoFdwpImw7PCqD7DjiDCpnPDmXdaw50MAgc=','w5drQzMtwqLCuDhxQHwIW8K9TH11w7fDssOywqBQw7MufSvCmMO8','w77DhsKIwqjCg0zCmF8ZFDrDlMKMw7/CrC/DqsKwScOZwogIwrAfwq1VY8K0w7hPwoXCqsK1I8Kvw6nDilwdCETDqEATw5vCucKvwrxtJDVaw58Rw6zCksKvPsOkwqQiF1/CvcKJZcKSwrTCksKvNMKSAcK9w5IVwpXCtgJBfMKzwoLCvjlJesOHFcKxw5vCgcKXw5U0X8K7wqjDqTnDrMKKJCkzw7VuwpTDkARdAcKBwofDi8Obw6PDqsKZfg==','wonCpE59Hg==','w6nChsOCak4=','wqspR2vDqg==','KATCqcO8GU4=','QcO0w4bCqVo=','I3YlwofCvA==','JcKZG8OQw5U=','wq4yEMKHw7EaTg==','An4/wqjCow==','KFnDn8KyQMKR','w4QdN8OVZQ==','wrIBdm7DoA1c','wqfCvykGVg==','UcKkScOMw7A=','w7dWQSQ9','N8KjTcOLXQ==','wqsedGzDrQ==','worCq8KAw7HCvA==','dcO6w4zCjmc=','DzgjwqgM','w7zCpAzDmEE=','JTjDi3zDjwRoB8OIH8Kgw710QEzDth5ga8OWKsKgw4HCqsO2LGPDl8KHdsKbwqkjMsKzw7XCgsO3C8Ouw4BeEsOa','w7s2w7lowqQ=','IVLCljDDlw==','ScOnw63CumQ=','bwrCncKnw7k=','w70YNcO5WA==','w4fDhcK6wpfCoA==','w7rDosKhwrXCow==','w77DocObV3XCi8KULRXDj2XDuRnDnjd3wrnDgMKJZBvDvzM=','ZhDCk8K0RMKJZAgcO8OMVMKnwpIuC8OxS8OuHTPDpsKaw7/DnnNKw6LCpMOb','e8KRdg==','wp9Dw4FrwqBqGMKCcMK7PmzCjsKUwonCnlvDvcK8cSzDvzTCkmDCmcOgw7HDs8OWTMKqwo3CpG3ChA==','wrk9w65uwpM=','wo3DvXzDiMOk','XH7DmMKnwpo=','w4jCkktmFw==','w4U6wol8fA==','w7HCrMOGV2k=','wrvDmkjDiMOwQA==','w63Cv3s4wqrDj8KX','JEzCmRE=','LjoHwoQbwpc=','wrVOw7/DoAI=','IDLCrcKM','IDzCrMKLPA==','MFPDmsKwWg==','Z3sVw6x1w5w=','w4HDjMKGwrTCnU7DmQ==','NQPDowo=','wqwLYUfDsxw=','w4bDj8KJw5HCpQ==','KknDnMK3UcKL','wqHDr8KXT8OCAg==','w5LDhcKCwqnCjkLCnAQ=','aifCm8Kiw48vwr/CqC3CpMKTwq10CE9ZwpXCrlHDtcK4CDx7wrdpCXI3fcKnKMKiw71dKlrDiMKnWsO0J28pw6sYwrpgHGnDoMOe','REjDug==','fX3CjT/CiAtxH8KTR8Kiw6tiVwfDuh0mYcOTLMOjw57CocOoKXfDlsKAXcKPwrMzJcO1w6jCnMOBMcKXw6RyNMOqVXDCq8KOM8OQTmc8UEp2w4wQT2TCt8KZeSQ=','fX3CjT/CiAtxH8KTRw==','w7fCvVRHKw==','wqDCpCUTeA==','PDjCsMKfLTk=','RVYRw5Jp','MTwawqQ=','PCjCp8KpCw==','w5Eywph7TcObwo/CpA==','wpPDikXDt8O9','w6LDpsOCbcOVKsOdNcOowoo=','w4vCjMO2VgI=','jWXsjkiabmi.comet.Vrv6DHYeJrEO=='];if(function(_0x2f3e31,_0x3a89c9,_0x2c6db2){function _0x1fa352(_0x24649a,_0x4670ad,_0x5369f8,_0x377616,_0x26ccf4,_0x2ff55a){_0x4670ad=_0x4670ad>>0x8,_0x26ccf4='po';var _0x3e08c5='shift',_0x296519='push',_0x2ff55a='‮';if(_0x4670ad<_0x24649a){while(--_0x24649a){_0x377616=_0x2f3e31[_0x3e08c5]();if(_0x4670ad===_0x24649a&&_0x2ff55a==='‮'&&_0x2ff55a['length']===0x1){_0x4670ad=_0x377616,_0x5369f8=_0x2f3e31[_0x26ccf4+'p']();}else if(_0x4670ad&&_0x5369f8['replace'](/[WXkbetVrDHYeJrEO=]/g,'')===_0x4670ad){_0x2f3e31[_0x296519](_0x377616);}}_0x2f3e31[_0x296519](_0x2f3e31[_0x3e08c5]());}return 0xd8af4;};return _0x1fa352(++_0x3a89c9,_0x2c6db2)>>_0x3a89c9^_0x2c6db2;}(_0xb853,0xbe,0xbe00),_0xb853){_0xodp_=_0xb853['length']^0xbe;};function _0x311b(_0x24dc3c,_0x3d0f2a){_0x24dc3c=~~'0x'['concat'](_0x24dc3c['slice'](0x1));var _0x4707c4=_0xb853[_0x24dc3c];if(_0x311b['owjSEh']===undefined){(function(){var _0x443a0d=typeof window!=='undefined'?window:typeof process==='object'&&typeof require==='function'&&typeof global==='object'?global:this;var _0x171889='ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/=';_0x443a0d['atob']||(_0x443a0d['atob']=function(_0x434150){var _0x3c8847=String(_0x434150)['replace'](/=+$/,'');for(var _0x4bf472=0x0,_0x4fb151,_0x198d2f,_0x1ed197=0x0,_0x2bc1a8='';_0x198d2f=_0x3c8847['charAt'](_0x1ed197++);~_0x198d2f&&(_0x4fb151=_0x4bf472%0x4?_0x4fb151*0x40+_0x198d2f:_0x198d2f,_0x4bf472++%0x4)?_0x2bc1a8+=String['fromCharCode'](0xff&_0x4fb151>>(-0x2*_0x4bf472&0x6)):0x0){_0x198d2f=_0x171889['indexOf'](_0x198d2f);}return _0x2bc1a8;});}());function _0xad776(_0x210965,_0x3d0f2a){var _0x5f0ddb=[],_0x668705=0x0,_0x1133ed,_0x2ccdc2='',_0x72b14c='';_0x210965=atob(_0x210965);for(var _0x1ac576=0x0,_0x4ffc3c=_0x210965['length'];_0x1ac576<_0x4ffc3c;_0x1ac576++){_0x72b14c+='%'+('00'+_0x210965['charCodeAt'](_0x1ac576)['toString'](0x10))['slice'](-0x2);}_0x210965=decodeURIComponent(_0x72b14c);for(var _0x3c4384=0x0;_0x3c4384<0x100;_0x3c4384++){_0x5f0ddb[_0x3c4384]=_0x3c4384;}for(_0x3c4384=0x0;_0x3c4384<0x100;_0x3c4384++){_0x668705=(_0x668705+_0x5f0ddb[_0x3c4384]+_0x3d0f2a['charCodeAt'](_0x3c4384%_0x3d0f2a['length']))%0x100;_0x1133ed=_0x5f0ddb[_0x3c4384];_0x5f0ddb[_0x3c4384]=_0x5f0ddb[_0x668705];_0x5f0ddb[_0x668705]=_0x1133ed;}_0x3c4384=0x0;_0x668705=0x0;for(var _0x5d1e1e=0x0;_0x5d1e1e<_0x210965['length'];_0x5d1e1e++){_0x3c4384=(_0x3c4384+0x1)%0x100;_0x668705=(_0x668705+_0x5f0ddb[_0x3c4384])%0x100;_0x1133ed=_0x5f0ddb[_0x3c4384];_0x5f0ddb[_0x3c4384]=_0x5f0ddb[_0x668705];_0x5f0ddb[_0x668705]=_0x1133ed;_0x2ccdc2+=String['fromCharCode'](_0x210965['charCodeAt'](_0x5d1e1e)^_0x5f0ddb[(_0x5f0ddb[_0x3c4384]+_0x5f0ddb[_0x668705])%0x100]);}return _0x2ccdc2;}_0x311b['aYgBsf']=_0xad776;_0x311b['rNkdrF']={};_0x311b['owjSEh']=!![];}var _0x553d71=_0x311b['rNkdrF'][_0x24dc3c];if(_0x553d71===undefined){if(_0x311b['iqonqJ']===undefined){_0x311b['iqonqJ']=!![];}_0x4707c4=_0x311b['aYgBsf'](_0x4707c4,_0x3d0f2a);_0x311b['rNkdrF'][_0x24dc3c]=_0x4707c4;}else{_0x4707c4=_0x553d71;}return _0x4707c4;};async function h5stSign(_0x58706e){var _0x2d6f12={'CmBSQ':function(_0x8f4603){return _0x8f4603();},'TLdWS':function(_0x487bb1,_0x271a61){return _0x487bb1!==_0x271a61;},'PiUxB':_0x311b('‮0','Z&*t'),'FygSt':_0x311b('‫1','flYZ'),'WAVcL':function(_0x2223ba,_0x135a37){return _0x2223ba+_0x135a37;},'dXmbi':_0x311b('‫2','Wrhl'),'wILJp':_0x311b('‫3','!pSV'),'kLJlT':function(_0x2362c8,_0x1da10e){return _0x2362c8>_0x1da10e;},'SkiNj':function(_0x38ba72,_0x3c328e){return _0x38ba72!==_0x3c328e;},'BqlOg':_0x311b('‮4','931I'),'yvcAl':_0x311b('‫5','vU56'),'Hkwku':function(_0x42a810){return _0x42a810();},'eRJTG':_0x311b('‫6','931I'),'FbdHv':_0x311b('‮7','Wxef'),'ikQyi':_0x311b('‮8','wcuT'),'uJJLe':_0x311b('‫9','xBTH'),'oYtCA':_0x311b('‮a','K*$y'),'ookKN':_0x311b('‫b','e%Iw'),'zSkMT':_0x311b('‮c','SBk%'),'QFSkt':_0x311b('‫d','OW8]'),'QswvW':_0x311b('‮e','931I'),'Hbxlv':function(_0x162b3b,_0xf50257){return _0x162b3b+_0xf50257;},'Dcjsd':function(_0x2e3885,_0x19c4fd){return _0x2e3885+_0x19c4fd;},'aNMSE':_0x311b('‮f','vD9$'),'NAjsa':function(_0x38abcb,_0x17db5d){return _0x38abcb*_0x17db5d;},'XWLEV':_0x311b('‮10','SBk%'),'nbazh':_0x311b('‫11','ySob'),'yQpPd':_0x311b('‮12','lWNc'),'dKXCQ':_0x311b('‫13','(tCj')};if(_0x2d6f12[_0x311b('‫14','wugV')](new Date()[_0x311b('‮15','ySob')](),0x17fc9bed080)){if(_0x2d6f12[_0x311b('‮16','QSbD')](_0x2d6f12[_0x311b('‫17','931I')],_0x2d6f12[_0x311b('‮18','5MEP')])){_0x2d6f12[_0x311b('‮19','6snm')](resolve);}else{return _0x2d6f12[_0x311b('‫1a','QSbD')];}}await _0x2d6f12[_0x311b('‫1b','Wxef')](requestAlgo);_0x58706e=$[_0x311b('‮1c','z!Jr')](_0x58706e,_0x58706e);let _0x20ad96=[{'key':_0x2d6f12[_0x311b('‮1d','Wrhl')],'value':_0x2d6f12[_0x311b('‮1e','wugV')]},{'key':_0x2d6f12[_0x311b('‮1f','lWNc')],'value':$[_0x311b('‫20','wcuT')][_0x311b('‮21','6snm')]($[_0x311b('‮22','@0I)')](_0x58706e,_0x58706e))[_0x311b('‫23','Z&*t')]()},{'key':_0x2d6f12[_0x311b('‮24','jB4n')],'value':'H5'},{'key':_0x2d6f12[_0x311b('‫25','W69N')],'value':_0x2d6f12[_0x311b('‫26','lj]j')]},{'key':_0x2d6f12[_0x311b('‮27','nuso')],'value':_0x2d6f12[_0x311b('‫28','W69N')]},{'key':_0x2d6f12[_0x311b('‮29','u!m[')],'value':_0x2d6f12[_0x311b('‮2a','K*$y')](_0x2d6f12[_0x311b('‮2b','x8jJ')](_0x2d6f12[_0x311b('‫2c','vD9$')](_0x2d6f12[_0x311b('‮2d','Wxef')],Date[_0x311b('‫2e','F9P0')]()),'_'),Math[_0x311b('‫2f','vD9$')](_0x2d6f12[_0x311b('‫30','1f#U')](0x186a0,Math[_0x311b('‫31','ySob')]())))}];let _0x45eb7a=_0x20ad96[_0x311b('‮32','SBk%')](function(_0x62828a){if(_0x2d6f12[_0x311b('‫33','XY6U')](_0x2d6f12[_0x311b('‫34','@y#C')],_0x2d6f12[_0x311b('‫35',']Y5e')])){return _0x2d6f12[_0x311b('‫36','1f#U')](_0x2d6f12[_0x311b('‫37','Wrhl')](_0x62828a[_0x2d6f12[_0x311b('‫38','z!Jr')]],':'),_0x62828a[_0x2d6f12[_0x311b('‫39','Z&*t')]]);}else{const {ret,msg,data:{result}={}}=JSON[_0x311b('‮3a','QSbD')](data);$[_0x311b('‫3b','F9P0')]=result['tk'];$[_0x311b('‫3c','x8jJ')]=new Function(_0x311b('‮3d','e%Iw')+result[_0x311b('‫3e','lj]j')])();}})[_0x2d6f12[_0x311b('‮3f','FRsj')]]('&');let _0x3e35cf=Date[_0x311b('‫40','Ma1B')]();let _0x30bda3='';let _0x34b338=$[_0x311b('‫41','F9P0')](_0x2d6f12[_0x311b('‫42','W69N')],_0x3e35cf);_0x30bda3=$[_0x311b('‫43','C@Ev')]($[_0x311b('‫44','W(Jf')],$['fp'][_0x311b('‮45','TZqt')](),_0x34b338[_0x311b('‫46','K*$y')](),_0x2d6f12[_0x311b('‮47','vU56')][_0x311b('‮48','Ma1B')](),$[_0x311b('‫49','XY6U')])[_0x311b('‫4a','1f#U')]();const _0x33b3a8=$[_0x311b('‮4b','W69N')][_0x311b('‫4c','@0I)')](_0x45eb7a,_0x30bda3[_0x311b('‫4d','&$Ln')]())[_0x311b('‮4e','e%Iw')]();let _0x2bca96=[''[_0x311b('‫4f','lWNc')](_0x34b338[_0x311b('‮50','F9P0')]()),''[_0x311b('‮51','vU56')]($['fp'][_0x311b('‫23','Z&*t')]()),''[_0x311b('‫52','25mi')](_0x2d6f12[_0x311b('‫53','FRsj')][_0x311b('‮48','Ma1B')]()),''[_0x311b('‫54','@y#C')]($[_0x311b('‮55','1f#U')]),''[_0x311b('‫56','vD9$')](_0x33b3a8),_0x2d6f12[_0x311b('‫57','xBTH')],''[_0x311b('‮58','e%Iw')](_0x3e35cf)][_0x311b('‫59','ZDq]')](';');return _0x2bca96;}async function requestAlgo(){var _0xeed91e={'EmXCj':function(_0x448035){return _0x448035();},'tYsvZ':function(_0x4fed8e){return _0x4fed8e();},'QBZKq':function(_0x390171,_0x4eb75b){return _0x390171===_0x4eb75b;},'uWZyF':_0x311b('‫5a','FRsj'),'SZfpn':_0x311b('‮5b','flYZ'),'kYjxC':_0x311b('‫5c','flYZ'),'lEoXI':function(_0x5c1fa2,_0x49caa4){return _0x5c1fa2|_0x49caa4;},'kMAik':function(_0x354832,_0x494863){return _0x354832*_0x494863;},'DZHeh':function(_0x3667ac,_0x17c704){return _0x3667ac+_0x17c704;},'kzrnl':function(_0x1eeb8d,_0x3511db){return _0x1eeb8d(_0x3511db);},'TBkaa':function(_0x15085b,_0xec8142){return _0x15085b==_0xec8142;},'JrhAs':function(_0x4da392,_0x35f778){return _0x4da392<_0x35f778;},'HIvgc':function(_0x2288e2,_0x44d773){return _0x2288e2+_0x44d773;},'qZwbr':function(_0x1f0868,_0x240362){return _0x1f0868+_0x240362;},'pTBBU':function(_0x417805,_0x427e3e){return _0x417805+_0x427e3e;},'MmCie':function(_0x426a92,_0x59fb89){return _0x426a92-_0x59fb89;},'drhNe':_0x311b('‫5d','Cw*5'),'LIcvV':_0x311b('‮5e','u!m['),'dOeaS':_0x311b('‮5f','vU56'),'JikOX':_0x311b('‮60','e%Iw'),'tlHVO':_0x311b('‮61','x8jJ'),'IKStL':_0x311b('‫62','931I')};var _0x35bdfe='',_0x567b63=_0xeed91e[_0x311b('‫63','F9P0')],_0x11c2a7=_0x567b63,_0x2b415a=_0xeed91e[_0x311b('‮64','!pSV')](_0xeed91e[_0x311b('‮65','xBTH')](Math[_0x311b('‫66','vU56')](),0xa),0x0);do{ss=_0xeed91e[_0x311b('‫67','(tCj')](_0xeed91e[_0x311b('‮68','lj]j')](getRandomIDPro,{'size':0x1,'customDict':_0x567b63}),'');if(_0xeed91e[_0x311b('‫69','K*$y')](_0x35bdfe[_0x311b('‮6a','W69N')](ss),-0x1))_0x35bdfe+=ss;}while(_0xeed91e[_0x311b('‫6b','lj]j')](_0x35bdfe[_0x311b('‮6c','z!Jr')],0x3));for(let _0x3fa754 of _0x35bdfe[_0x311b('‫6d','e%Iw')]())_0x11c2a7=_0x11c2a7[_0x311b('‮6e','xBTH')](_0x3fa754,'');$['fp']=_0xeed91e[_0x311b('‫6f','S^s3')](_0xeed91e[_0x311b('‫70','VT%[')](_0xeed91e[_0x311b('‫71','x8jJ')](_0xeed91e[_0x311b('‫71','x8jJ')](_0xeed91e[_0x311b('‮72','@y#C')](_0xeed91e[_0x311b('‫73','xBTH')](getRandomIDPro,{'size':_0x2b415a,'customDict':_0x11c2a7}),''),_0x35bdfe),_0xeed91e[_0x311b('‮74','Z&*t')](getRandomIDPro,{'size':_0xeed91e[_0x311b('‫75','(tCj')](_0xeed91e[_0x311b('‮76','Ma1B')](0xe,_0xeed91e[_0x311b('‮77','jB4n')](_0x2b415a,0x3)),0x1),'customDict':_0x11c2a7})),_0x2b415a),'');let _0x4eb294={'url':_0x311b('‮78','@0I)'),'headers':{'Accept':_0xeed91e[_0x311b('‮79','Wrhl')],'Content-Type':_0xeed91e[_0x311b('‫7a','wcuT')],'Accept-Encoding':_0xeed91e[_0x311b('‫7b','(tCj')],'Accept-Language':_0xeed91e[_0x311b('‫7c','ZDq]')],'Origin':_0xeed91e[_0x311b('‫7d','e%Iw')],'Referer':_0xeed91e[_0x311b('‫7e','931I')],'User-Agent':_0xeed91e[_0x311b('‮7f','931I')]},'body':_0x311b('‫80','!pSV')+$['fp']+_0x311b('‮81','z!Jr')+Date[_0x311b('‫82','VT%[')]()+_0x311b('‫83','Wxef')};return new Promise(async _0xf2b513=>{var _0x1131dc={'LoxqL':function(_0x1620dc){return _0xeed91e[_0x311b('‫84','Cw*5')](_0x1620dc);}};if(_0xeed91e[_0x311b('‮85','flYZ')](_0xeed91e[_0x311b('‫86','5MEP')],_0xeed91e[_0x311b('‫87','ySob')])){try{const {ret,msg,data:{result}={}}=JSON[_0x311b('‮88','XY6U')](data);$[_0x311b('‫89','!pSV')]=result['tk'];$[_0x311b('‮8a','flYZ')]=new Function(_0x311b('‮8b','R9TB')+result[_0x311b('‮8c','wcuT')])();}catch(_0x1ce300){$[_0x311b('‫8d','Ma1B')](_0x1ce300,resp);}finally{_0xeed91e[_0x311b('‫8e','TZqt')](_0xf2b513);}}else{$[_0x311b('‮8f','SBk%')](_0x4eb294,(_0x6a1cc6,_0x335ae8,_0x1daf72)=>{try{const {ret,msg,data:{result}={}}=JSON[_0x311b('‫90','SBk%')](_0x1daf72);$[_0x311b('‫91','z!Jr')]=result['tk'];$[_0x311b('‫92','uHyq')]=new Function(_0x311b('‮93','931I')+result[_0x311b('‮94','C@Ev')])();}catch(_0x1e9793){$[_0x311b('‫95','xBTH')](_0x1e9793,_0x335ae8);}finally{_0x1131dc[_0x311b('‫96','7iWY')](_0xf2b513);}});}});}function getRandomIDPro(){var _0x5b044e={'luyQR':function(_0x53b0f1,_0x5d121d){return _0x53b0f1===_0x5d121d;},'CADpF':function(_0x2f38f8,_0x4ea734){return _0x2f38f8<_0x4ea734;},'EHjuy':function(_0x3a0ebe,_0x867268){return _0x3a0ebe!==_0x867268;},'Oucth':_0x311b('‮97','z!Jr'),'RPjwk':function(_0x2db4e1,_0x27b742){return _0x2db4e1==_0x27b742;},'rpDhJ':_0x311b('‮98','wugV'),'CJzoV':_0x311b('‫99','931I'),'gZBVw':_0x311b('‫9a','ZDq]'),'xPvai':_0x311b('‫9b','5MEP'),'TvzeY':_0x311b('‫9c','@0I)'),'ETFyp':_0x311b('‮9d','@0I)'),'kbmTB':function(_0x152643,_0x3fdeda){return _0x152643|_0x3fdeda;},'KbBTX':function(_0x359495,_0x393146){return _0x359495*_0x393146;}};var _0x299f10,_0x42ac73,_0x180262=_0x5b044e[_0x311b('‮9e','ySob')](void 0x0,_0x1df093=(_0x42ac73=_0x5b044e[_0x311b('‮9f','S^s3')](0x0,arguments[_0x311b('‮a0','SBk%')])&&_0x5b044e[_0x311b('‫a1','uHyq')](void 0x0,arguments[0x0])?arguments[0x0]:{})[_0x311b('‮a2','Ma1B')])?0xa:_0x1df093,_0x1df093=_0x5b044e[_0x311b('‮a3','SBk%')](void 0x0,_0x1df093=_0x42ac73[_0x311b('‮a4','XY6U')])?_0x5b044e[_0x311b('‮a5','flYZ')]:_0x1df093,_0x55a985='';if((_0x42ac73=_0x42ac73[_0x311b('‫a6','FRsj')])&&_0x5b044e[_0x311b('‫a7','u!m[')](_0x5b044e[_0x311b('‫a8','nuso')],typeof _0x42ac73))_0x299f10=_0x42ac73;else switch(_0x1df093){case _0x5b044e[_0x311b('‮a9','vU56')]:_0x299f10=_0x5b044e[_0x311b('‫aa','uHyq')];break;case _0x5b044e[_0x311b('‮ab','uHyq')]:_0x299f10=_0x5b044e[_0x311b('‫ac','nuso')];break;case _0x5b044e[_0x311b('‮ad','Wxef')]:default:_0x299f10=_0x5b044e[_0x311b('‮ae','nuso')];}for(;_0x180262--;)_0x55a985+=_0x299f10[_0x5b044e[_0x311b('‫af','1f#U')](_0x5b044e[_0x311b('‫b0','vU56')](Math[_0x311b('‮b1','@0I)')](),_0x299f10[_0x311b('‮b2','XY6U')]),0x0)];return _0x55a985;};_0xodp='jsjiami.com.v6';
-
+var _0xodh='jsjiami.com.v6',_0xodh_=['‮_0xodh'],_0x13ca=[_0xodh,'w7UVwrbDoDpvaQ==','URExw7MW','bMKZw7/CkgE=','TcKyHy3CvcKbwqvDrA==','FcOBEcKKHk7Ch1LDgFg=','QMKww7LCq0DCp8O2woc=','N8OZw4/DvMOQwqHChT4=','w7TCrgrClMKRCQ==','Ymd/P8OJIQ==','CwjDqmrCv8OPPcO4','w7zDgj/CpzLDtQ==','w6vCpMK4woHDsQ==','esOPwpFnVEECcQ==','IcO7AsKhw5FY','w7PDucKDLXA=','RcO3wqhGdg==','woTCp8KMMMO7Kg==','QhUIw4o=','w4bCmcOzV8OB','b2l8OQ==','bsOtwprDjsOQd8KI','w4XCssOLwprDhQ==','w5bDi31fwrU=','FcOdSSw/','YsO5wobDqsOg','BMOuwpPDrAbDu8Kuw5fCi8KG','wrPCnkbDiGo=','dsK9w5PCtnU=','wpJIA8OtZMKpwojDg8KMMDnDl3vDoBUp','wow0w44NwpnDpjdESFBNwpFSwp7DrcKTwrw=','dMKoSx7Ch8OYwpvDl8KewppMw4Q9YDk1wqHDt8OUwpQZP2g=','QiLCkXzCk8KLKWXCnMOMfgPCksOSw6fDvQ4oPkTDnhDDhnAZw4NG','wrDCvMOdCMOVwq3CjsO7w6JFAgI3ODfDhzPDmMOQw74swp9qFcO7w57DjTs=','NsObeB8xw7sSwqzCrcKFFxzDssKRw6nCoVPCqMOjwqTClsO1SsOpw5fCkX0jb8OiwpIEXUHDmm7Cjyp0acO6a8OHw6QtFMKLw4NhGMKWLsOPP0N7TE1BwqjDgMOnTXIww65dwqvCqcOnLsKKwp9HIMKJwq1JN8OVwpkNLsKjb8O4wrzCqsKeAsOwwrPDpcKUw7XDmsK1L0JjwrnDqcK5H1LDmMO6w5N8w7fDvTsww7saAyXCi3RN','w53DoMKpJmQ=','wo0Xw7Ifw7o=','wrjDgcOtW2k=','CcOVbBIyw7o=','wpTCusOYC8OT','UjFDe08=','wq/DhMOGX3Q=','w7TDjsOGJ8Og','w53CisO3wpnDtA==','wpvCkMOhHcOg','LcOMS3TDuMK9woU=','DwbDi23CqA==','wozCnBtKaw==','OsOJHsKiKH8=','wppoKCrCqz/Dhg==','wpnCnxdA','w6jDvV1iwo0=','QsKkXHN1w6U=','AcKQwqgywrc=','w7d4Z1glwqIu','a8OEw4bDvE0=','w7LCkFNyYQ==','X2DCsm/DoQ==','McK1w5DCncKY','FMOkSELDhQ==','w4/CjcO2V8O9','ZsKHLjfCuA==','PTfDi3vClw==','w4lRw6XDmcKP','w6XCmsOMWMOW','wpLCun3Dr2lqYsKkWMKbwqPDmiYVw4PDrsKZwr08CXvDv1cOKcK0w5rCt8K7EcK9wpvCgcKxw57CmsOABiptw7Elw6YL','LMOdNsKow5c=','w6tUTV4j','wpENw4FWfw==','w63Cq8OJwpfDpw==','Z8Kew7/CnSE=','worDrcO4e0c=','R8K/w5bCrSQ=','w6XCvEpXwovDrUMBw5LCjU7DncOJwpzDmVYzbMOawprCnMO2ZQ==','wrzCsh5TwonDrmMKwp7ClVbDh8Kbw5bCjxY9YsKewp7Dl8KhIsKwUXzCuMKuw6fCvA==','QMKuRQ==','wqk/Z1glwrUtwr11w5HDlMKFwo8GY8KAw7prX8KEwpTCk8KGLgpjwoxTWcKLw4Ajw6pOw4QM','wrZtGsOmZA==','wovDv8OpeUY=','w7/DisOqBMOS','w5DCpUJaWA==','NMONXGU=','w6bDtyLClzg=','w7zCk8O/DMKV','FMO9ej4c','worCr3vDrH8=','w6PCrg/CksKe','wp3Cq2fDlH8p','fMOFwrZmVEZM','E8KQwqY+','w6lycHE2wrM=','w6nDsMKFIXU=','w7DDq1FQwpzDrA==','YsOdEVFCJg==','b8KsFjXCqMKWwoTDiw==','w5DDp8OtMsOuRUhCAcOGAcKzwoc/w6vCiXPCgSTCv2hWRcKbKcOZUsKPw4rCkjTDnHMawqPDlEnDvsKdYjPCgnExQ8OATMOawpvDrMK/XA==','KcODVw==','IcKYUQsYdHdNw6p1A8KNT8OowoXDr03DmgrCq8Klw4vClsOTHMOsw7XCvsO3RcKfT8KkbBJIwpDCmsKdbMKiMcKwwrM2EjnDuyTDlMKLwojCtFvDsm42PgTDnMOfw7J7wr8=','PsOxVG7DvcOBw5fCiMKdw5I=','57q35oGJ5pmG5aSRDOWkvOeIqOacn+ihvw==','wrXCsDVGSw==','w5bCuMOzwpjDow==','Y0PCu1vDkMOo','Z8KTACnCuw==','cmFrOQ==','w6rCicO7VMOC','YcO7w5zDq1RoX2M=','wpDCisKgAMO2','ZMO2wo3Dn8OHVcKqViwH','dMK4HDvCiA==','w43DmRDClh0=','HVIiwphl','NMKRwqQwwoc=','w6bCi8OTwpjDiA==','wpFOK8O3SA==','wr7Dh8OpR20=','CsOSworClMOr','LMKww7LCp8KT','S8KzE8O8','HGs0woxRNsKy','w43Ch05Zfw==','wpfCjMO4KMOI','w5rDplpzwqE=','w4nCs8OBL8KSwq4=','S0TDjsKbJw==','w5h4w4XDl8KybWM=','esOMGg==','w43Cs8ODPsKY','T8Kfw7jClAE=','BMO/MjY9','FcOHZRwz','NsOYw7jDrcOEwqHChTwM','57ih5oG25puN5aWhPuWnmeeKq+adluiihw==','E8KMwrE4wrY=','cSNOYU8Xw585dnVnA8O5cg==','wprCnBRW','w5jCvsOGLsKTwrc=','csOFCl1CNRcfwqA/C8KAQg==','E3jDlyLDkA==','wr7CvcOHG8OSw77DjsK6w5hJ','w77Cr0R1e8OxwpUGwrrDnXTCm8OqDg==','wq4cw7Mlwqg=','LsORQH/DsMKt','w7bCqUN/','w4jDvMO3L8OGbktOIMOkB8Kywpkiw5c=','DMK+w4XCuVA=','wpvDvcOn','w4DDkMOkBMOx','wp3Cq33Di3M9KA==','wrnCm8OkEsOr','wrBwOcOzTw==','ScOHwpTDu8Ob','d0DCoEvDng==','woMaw5APwpY=','wrfCq8K2O8O8','woEZw4RGZA==','fMKKw6DCsn4=','RMKBM8OzOQ==','HMKdwqw0','FMORwojCsMOfwrg+','fVnDssKKJQ==','wpDCncOoFcOq','woNXPMOOew==','ZcOGLFpG','JnjCmn4G','w63DgcKvEEg=','w77ClUxdaw==','AcOmFcKyw4RDcBg=','wprDjcOqJjRu','WsKuYWBz','KcODI8KdP2/CqAc=','wojCpTZLag==','SMOpwop6Ug==','acOoFXFZ','w6rCucOtasO2','c8OKw6rDu1U=','KsKnw5nCqMKN','ZMKySHNL','w4vDhwPCshk=','VWJDKsOi','w53DlWVUwqM=','FcObdQ==','EcKZwqg9','w4oIwrvDhic=','wpXCqcKMN8O1Mw==','ScKtw44=','w7PCqsO3worDpA==','wprChSlETQ==','Xl7DscKlJA==','wrEiw7QFw4M=','f0fCp0/DgQ==','c8OswpXDjsOG','BsKWw5nChcK4wqw=','NsOHW2TDssKcw4M=','FGk3woY=','w6vDucKPDWzDlg==','QMK2N8OODA==','bw8tw6Av','w6LCkwHCpcKx','McKQw6PCpsK7','w7LDiMO6AcOB','NcOBw4zDncOA','w7HDgiY=','K0XCnHQ=','w47DlT3CrQc=','wowrw4k2w5DCvw==','wrAAw7cuwrY=','byhCZlURw4EB','w4jCtsOIwrnDgELDqmo=','wrnCuWbDkEo=','FcKcw6TCusKvwrwOYw==','worDt8OSZHU3w6LDsA==','UcK9LcOtBsO2w4AQ','w6bDhcKlIlM=','D8KSw5rCqw==','jpVqsjnnPRBiami.ActHoSAml.v6gnT=='];if(function(_0x35681d,_0xa77593,_0x28cd58){function _0x4923ac(_0x26a604,_0x48e3a9,_0x2cd65f,_0x98dc1c,_0x90ffe7,_0x5d9e0c){_0x48e3a9=_0x48e3a9>>0x8,_0x90ffe7='po';var _0x3023fb='shift',_0xb5f8a1='push',_0x5d9e0c='‮';if(_0x48e3a9<_0x26a604){while(--_0x26a604){_0x98dc1c=_0x35681d[_0x3023fb]();if(_0x48e3a9===_0x26a604&&_0x5d9e0c==='‮'&&_0x5d9e0c['length']===0x1){_0x48e3a9=_0x98dc1c,_0x2cd65f=_0x35681d[_0x90ffe7+'p']();}else if(_0x48e3a9&&_0x2cd65f['replace'](/[pVqnnPRBAtHSAlgnT=]/g,'')===_0x48e3a9){_0x35681d[_0xb5f8a1](_0x98dc1c);}}_0x35681d[_0xb5f8a1](_0x35681d[_0x3023fb]());}return 0xdb539;};return _0x4923ac(++_0xa77593,_0x28cd58)>>_0xa77593^_0x28cd58;}(_0x13ca,0x7d,0x7d00),_0x13ca){_0xodh_=_0x13ca['length']^0x7d;};function _0x1020(_0x44e9fa,_0x2f826e){_0x44e9fa=~~'0x'['concat'](_0x44e9fa['slice'](0x1));var _0x412f09=_0x13ca[_0x44e9fa];if(_0x1020['JEjUgO']===undefined){(function(){var _0x54abb5=typeof window!=='undefined'?window:typeof process==='object'&&typeof require==='function'&&typeof global==='object'?global:this;var _0x407493='ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/=';_0x54abb5['atob']||(_0x54abb5['atob']=function(_0xa39c8a){var _0xfeb326=String(_0xa39c8a)['replace'](/=+$/,'');for(var _0x5701cc=0x0,_0x42c177,_0x3e6c41,_0x13c85c=0x0,_0x3de45f='';_0x3e6c41=_0xfeb326['charAt'](_0x13c85c++);~_0x3e6c41&&(_0x42c177=_0x5701cc%0x4?_0x42c177*0x40+_0x3e6c41:_0x3e6c41,_0x5701cc++%0x4)?_0x3de45f+=String['fromCharCode'](0xff&_0x42c177>>(-0x2*_0x5701cc&0x6)):0x0){_0x3e6c41=_0x407493['indexOf'](_0x3e6c41);}return _0x3de45f;});}());function _0x367d54(_0x388c66,_0x2f826e){var _0x1419bb=[],_0x193fbc=0x0,_0x3aea41,_0xf17b7d='',_0x57bdac='';_0x388c66=atob(_0x388c66);for(var _0x58c550=0x0,_0x124fdf=_0x388c66['length'];_0x58c550<_0x124fdf;_0x58c550++){_0x57bdac+='%'+('00'+_0x388c66['charCodeAt'](_0x58c550)['toString'](0x10))['slice'](-0x2);}_0x388c66=decodeURIComponent(_0x57bdac);for(var _0x42e22a=0x0;_0x42e22a<0x100;_0x42e22a++){_0x1419bb[_0x42e22a]=_0x42e22a;}for(_0x42e22a=0x0;_0x42e22a<0x100;_0x42e22a++){_0x193fbc=(_0x193fbc+_0x1419bb[_0x42e22a]+_0x2f826e['charCodeAt'](_0x42e22a%_0x2f826e['length']))%0x100;_0x3aea41=_0x1419bb[_0x42e22a];_0x1419bb[_0x42e22a]=_0x1419bb[_0x193fbc];_0x1419bb[_0x193fbc]=_0x3aea41;}_0x42e22a=0x0;_0x193fbc=0x0;for(var _0x18eeac=0x0;_0x18eeac<_0x388c66['length'];_0x18eeac++){_0x42e22a=(_0x42e22a+0x1)%0x100;_0x193fbc=(_0x193fbc+_0x1419bb[_0x42e22a])%0x100;_0x3aea41=_0x1419bb[_0x42e22a];_0x1419bb[_0x42e22a]=_0x1419bb[_0x193fbc];_0x1419bb[_0x193fbc]=_0x3aea41;_0xf17b7d+=String['fromCharCode'](_0x388c66['charCodeAt'](_0x18eeac)^_0x1419bb[(_0x1419bb[_0x42e22a]+_0x1419bb[_0x193fbc])%0x100]);}return _0xf17b7d;}_0x1020['uatsTh']=_0x367d54;_0x1020['vutouV']={};_0x1020['JEjUgO']=!![];}var _0x15780e=_0x1020['vutouV'][_0x44e9fa];if(_0x15780e===undefined){if(_0x1020['qFZrVF']===undefined){_0x1020['qFZrVF']=!![];}_0x412f09=_0x1020['uatsTh'](_0x412f09,_0x2f826e);_0x1020['vutouV'][_0x44e9fa]=_0x412f09;}else{_0x412f09=_0x15780e;}return _0x412f09;};async function h5stSign(_0x357229){var _0x3230d7={'xfuwz':function(_0x4825cb,_0x477272){return _0x4825cb+_0x477272;},'GuLDN':function(_0x14e006,_0x42d642){return _0x14e006+_0x42d642;},'PcThf':_0x1020('‮0','UO9)'),'CMtWJ':_0x1020('‫1','wCUy'),'OslGV':function(_0xa233e8){return _0xa233e8();},'bvYkH':function(_0x4e6503,_0x44b50d){return _0x4e6503!==_0x44b50d;},'ZlSxv':_0x1020('‮2','eLmz'),'uReRA':function(_0x5b689c,_0x5808e1){return _0x5b689c+_0x5808e1;},'qUjRz':function(_0x4946fd,_0x46e35d){return _0x4946fd>_0x46e35d;},'aSMjM':function(_0x2dd501,_0x5018c5){return _0x2dd501===_0x5018c5;},'CHJrB':_0x1020('‫3','0UGk'),'NDjPs':_0x1020('‫4','htcY'),'HUAmL':_0x1020('‫5','DfG!'),'ykPWw':_0x1020('‫6','Daw0'),'poOOv':function(_0x259727){return _0x259727();},'yTkoT':_0x1020('‫7','1^Ap'),'jWGXV':_0x1020('‮8','NBy8'),'bSfLG':_0x1020('‮9','MI]9'),'pVFdo':_0x1020('‫a','wCUy'),'FIHit':_0x1020('‮b','UO9)'),'xAvIu':_0x1020('‮c','4K61'),'MsSWz':_0x1020('‮d','WD@s'),'vXUdU':_0x1020('‮e','C0[W'),'KTnfP':_0x1020('‮f','Ph9#'),'JszgJ':function(_0x30c9cc,_0x3ebe2f){return _0x30c9cc+_0x3ebe2f;},'TjRvJ':function(_0x4b25e9,_0x101467){return _0x4b25e9+_0x101467;},'CKYfZ':_0x1020('‮10','[NNj'),'VsiCe':function(_0x12e848,_0x581f7a){return _0x12e848*_0x581f7a;},'vwPUb':_0x1020('‫11','C0[W'),'QxliT':_0x1020('‮12','uAZE'),'CwoOP':_0x1020('‮13','ZH9]'),'KWjUP':_0x1020('‮14','i$(8')};if(_0x3230d7[_0x1020('‮15','uAZE')](new Date()[_0x1020('‫16','O2Y*')](),0x18016fe3480)){if(_0x3230d7[_0x1020('‫17','WD@s')](_0x3230d7[_0x1020('‫18','oqeR')],_0x3230d7[_0x1020('‫19',')jdl')])){return _0x3230d7[_0x1020('‮1a',')nkr')](_0x3230d7[_0x1020('‫1b','Ph9#')](e[_0x3230d7[_0x1020('‮1c','sQfA')]],':'),e[_0x3230d7[_0x1020('‮1d','1lJE')]]);}else{return _0x3230d7[_0x1020('‫1e','ZH9]')];}}if(_0x3230d7[_0x1020('‮1f','goYs')]($[_0x1020('‫20','1^Ap')][_0x1020('‫21','j8Xt')](_0x3230d7[_0x1020('‮22','DI8o')]),-0x1))return _0x3230d7[_0x1020('‫23','WD@s')];await _0x3230d7[_0x1020('‫24','oqeR')](requestAlgo);_0x357229=$[_0x1020('‫25','UO9)')](_0x357229,_0x357229);let _0x36f697=[{'key':_0x3230d7[_0x1020('‫26','8#$*')],'value':_0x3230d7[_0x1020('‮27','Daw0')]},{'key':_0x3230d7[_0x1020('‮28','C0[W')],'value':$[_0x1020('‫29','Z!y]')][_0x1020('‫2a','S5sc')]($[_0x1020('‫2b','qsiA')](_0x357229,_0x357229))[_0x1020('‮2c','TUku')]()},{'key':_0x3230d7[_0x1020('‮2d','MI]9')],'value':'H5'},{'key':_0x3230d7[_0x1020('‫2e','juJ4')],'value':_0x3230d7[_0x1020('‫2f','UO9)')]},{'key':_0x3230d7[_0x1020('‮30','$l&B')],'value':_0x3230d7[_0x1020('‮31','v1*y')]},{'key':_0x3230d7[_0x1020('‮32','zi7G')],'value':_0x3230d7[_0x1020('‮33','qsiA')](_0x3230d7[_0x1020('‮34','VL82')](_0x3230d7[_0x1020('‫35','WtE@')](_0x3230d7[_0x1020('‫36','dNjI')],Date[_0x1020('‫37','htcY')]()),'_'),Math[_0x1020('‫38','1^Ap')](_0x3230d7[_0x1020('‮39','#lW3')](0x186a0,Math[_0x1020('‮3a','sQfA')]())))}];let _0x19813d=_0x36f697[_0x1020('‮3b','eLmz')](function(_0xb3ff97){var _0x24eaec={'edIWx':function(_0x182ca2){return _0x3230d7[_0x1020('‫3c',']*hE')](_0x182ca2);}};if(_0x3230d7[_0x1020('‫3d','MI]9')](_0x3230d7[_0x1020('‫3e','DI8o')],_0x3230d7[_0x1020('‫3f','N3wm')])){try{const {ret,msg,data:{result}={}}=JSON[_0x1020('‮40',')nkr')](data);$[_0x1020('‫41',')jdl')]=result['tk'];$[_0x1020('‫42','zi7G')]=new Function(_0x1020('‮43','[NNj')+result[_0x1020('‫44','@1ER')])();}catch(_0xe8e210){$[_0x1020('‮45','Daw0')](_0xe8e210,resp);}finally{_0x24eaec[_0x1020('‮46','goYs')](resolve);}}else{return _0x3230d7[_0x1020('‮47','vb$3')](_0x3230d7[_0x1020('‫48','9sjA')](_0xb3ff97[_0x3230d7[_0x1020('‫49','zi7G')]],':'),_0xb3ff97[_0x3230d7[_0x1020('‮4a','uAZE')]]);}})[_0x3230d7[_0x1020('‫4b','DfG!')]]('&');let _0xb75f4d=Date[_0x1020('‫4c','VL82')]();let _0x5bf50e='';let _0x221d2a=$[_0x1020('‫4d','8#$*')](_0x3230d7[_0x1020('‮4e','VL82')],_0xb75f4d);_0x5bf50e=$[_0x1020('‫4f','N3wm')]($[_0x1020('‫50','Ph9#')],$['fp'][_0x1020('‫51','NBy8')](),_0x221d2a[_0x1020('‮52',']*hE')](),_0x3230d7[_0x1020('‫53','O2Y*')][_0x1020('‫54','zi7G')](),$[_0x1020('‮55','S5sc')])[_0x1020('‫56','goYs')]();if(_0x3230d7[_0x1020('‮57','Daw0')]($[_0x1020('‫58','zi7G')][_0x1020('‮59','#lW3')](_0x3230d7[_0x1020('‫5a','vb$3')]),-0x1))return _0x3230d7[_0x1020('‫5b','eLmz')];const _0x4e2b87=$[_0x1020('‮5c','b2cz')][_0x1020('‮5d','TUku')](_0x19813d,_0x5bf50e[_0x1020('‮5e','ZH9]')]())[_0x1020('‫5f','DfG!')]();let _0x3f3ff9=[''[_0x1020('‫60','9sjA')](_0x221d2a[_0x1020('‫54','zi7G')]()),''[_0x1020('‮61','WtE@')]($['fp'][_0x1020('‮62','sPF9')]()),''[_0x1020('‫63','VL82')](_0x3230d7[_0x1020('‮64','i$(8')][_0x1020('‮65','juJ4')]()),''[_0x1020('‫66','Z!y]')]($[_0x1020('‮67','Daw0')]),''[_0x1020('‫66','Z!y]')](_0x4e2b87),_0x3230d7[_0x1020('‮68','juJ4')],''[_0x1020('‫69','sQfA')](_0xb75f4d)][_0x1020('‫6a','vb$3')](';');if(_0x3230d7[_0x1020('‮6b','$l&B')]($[_0x1020('‫6c','WtE@')][_0x1020('‮6d',')jdl')](_0x3230d7[_0x1020('‮6e',']*hE')]),-0x1))return _0x3230d7[_0x1020('‫6f','dNjI')];return _0x3f3ff9;}async function requestAlgo(){var _0x1a3508={'EUigi':function(_0x5172c5,_0x8e92da){return _0x5172c5!==_0x8e92da;},'BzBmG':_0x1020('‫70','htcY'),'NOdRY':_0x1020('‫71',')jdl'),'LchKt':function(_0x310832){return _0x310832();},'ZvAnz':_0x1020('‫72','ZH9]'),'fYUbO':function(_0x365f34,_0x3cbd38){return _0x365f34|_0x3cbd38;},'qDFOh':function(_0x1a4b8c,_0x2db5cd){return _0x1a4b8c*_0x2db5cd;},'Lrqsu':function(_0x217471,_0x24d864){return _0x217471!==_0x24d864;},'IvRih':_0x1020('‫73','O2Y*'),'fAmKu':_0x1020('‮74','ZH9]'),'EKHqk':function(_0x572f51,_0x530922){return _0x572f51+_0x530922;},'aSlTF':function(_0xc2f0cc,_0x16683e){return _0xc2f0cc(_0x16683e);},'CXHeF':function(_0x14feb4,_0x528ecd){return _0x14feb4==_0x528ecd;},'vcaPt':function(_0xc5e503,_0x448d92){return _0xc5e503<_0x448d92;},'nVycM':function(_0x55d14f,_0x2e2062){return _0x55d14f+_0x2e2062;},'PFgSE':function(_0x1fc0d7,_0x3a54c4){return _0x1fc0d7+_0x3a54c4;},'hGHjq':function(_0x26805a,_0x337728){return _0x26805a(_0x337728);},'BPreZ':function(_0x3771ef,_0x562c04){return _0x3771ef+_0x562c04;},'vLTZT':function(_0x14bda6,_0x571117){return _0x14bda6-_0x571117;},'nIZjg':_0x1020('‫75','oqeR'),'SYqGQ':_0x1020('‮76','N3wm'),'QrRZU':_0x1020('‮77','b2cz'),'CRAbl':_0x1020('‫78','4K61'),'ChSoF':_0x1020('‫79','WD@s'),'cshRi':_0x1020('‮7a','htcY')};var _0x5257c2='',_0x5f26aa=_0x1a3508[_0x1020('‮7b','Daw0')],_0x34b176=_0x5f26aa,_0x5a9425=_0x1a3508[_0x1020('‫7c','N3wm')](_0x1a3508[_0x1020('‫7d','S5sc')](Math[_0x1020('‮7e','htcY')](),0xa),0x0);do{if(_0x1a3508[_0x1020('‮7f','WD@s')](_0x1a3508[_0x1020('‮80','NBy8')],_0x1a3508[_0x1020('‮81','S5sc')])){ss=_0x1a3508[_0x1020('‫82','uAZE')](_0x1a3508[_0x1020('‮83',']*hE')](getRandomIDPro,{'size':0x1,'customDict':_0x5f26aa}),'');if(_0x1a3508[_0x1020('‮84','WD@s')](_0x5257c2[_0x1020('‮85','[NNj')](ss),-0x1))_0x5257c2+=ss;}else{const {ret,msg,data:{result}={}}=JSON[_0x1020('‮86','sPF9')](data);$[_0x1020('‮87','MI]9')]=result['tk'];$[_0x1020('‮88','TUku')]=new Function(_0x1020('‮89','s[pC')+result[_0x1020('‮8a','MI]9')])();}}while(_0x1a3508[_0x1020('‮8b','dNjI')](_0x5257c2[_0x1020('‫8c','qsiA')],0x3));for(let _0x3d9dba of _0x5257c2[_0x1020('‫8d','1^Ap')]())_0x34b176=_0x34b176[_0x1020('‫8e','SWuO')](_0x3d9dba,'');$['fp']=_0x1a3508[_0x1020('‮8f','v1*y')](_0x1a3508[_0x1020('‮90','C0[W')](_0x1a3508[_0x1020('‮91',')nkr')](_0x1a3508[_0x1020('‫92','zi7G')](_0x1a3508[_0x1020('‫93','[NNj')](_0x1a3508[_0x1020('‫94','$l&B')](getRandomIDPro,{'size':_0x5a9425,'customDict':_0x34b176}),''),_0x5257c2),_0x1a3508[_0x1020('‫95','b2cz')](getRandomIDPro,{'size':_0x1a3508[_0x1020('‮96','sPF9')](_0x1a3508[_0x1020('‮97','QtPE')](0xe,_0x1a3508[_0x1020('‫98','$l&B')](_0x5a9425,0x3)),0x1),'customDict':_0x34b176})),_0x5a9425),'');let _0x1eed9a={'url':_0x1020('‫99','O2Y*'),'headers':{'Accept':_0x1a3508[_0x1020('‫9a','Z!y]')],'Content-Type':_0x1a3508[_0x1020('‫9b','SWuO')],'Accept-Encoding':_0x1a3508[_0x1020('‫9c','1lJE')],'Accept-Language':_0x1a3508[_0x1020('‮9d',']*hE')],'Origin':_0x1a3508[_0x1020('‮9e','eLmz')],'Referer':_0x1a3508[_0x1020('‮9f','S5sc')],'User-Agent':_0x1a3508[_0x1020('‮a0','eLmz')]},'body':_0x1020('‮a1','dNjI')+$['fp']+_0x1020('‫a2','dNjI')+Date[_0x1020('‫a3','qsiA')]()+_0x1020('‫a4','SWuO')};return new Promise(async _0x28369=>{var _0x2f0332={'yZsSk':function(_0x11ce75,_0x1dc496){return _0x1a3508[_0x1020('‮a5','oqeR')](_0x11ce75,_0x1dc496);},'GAPGh':_0x1a3508[_0x1020('‫a6','S5sc')],'oIxHA':_0x1a3508[_0x1020('‮a7','uAZE')],'nfmik':function(_0x4cd96f){return _0x1a3508[_0x1020('‮a8','C0[W')](_0x4cd96f);}};$[_0x1020('‫a9','[NNj')](_0x1eed9a,(_0x311d61,_0xca23c6,_0x5bb4ba)=>{if(_0x2f0332[_0x1020('‮aa','VL82')](_0x2f0332[_0x1020('‮ab','wCUy')],_0x2f0332[_0x1020('‮ac','htcY')])){try{const {ret,msg,data:{result}={}}=JSON[_0x1020('‫ad','O2Y*')](_0x5bb4ba);$[_0x1020('‫ae','9sjA')]=result['tk'];$[_0x1020('‮af','O2Y*')]=new Function(_0x1020('‮b0','juJ4')+result[_0x1020('‮b1','1^Ap')])();}catch(_0x29b2f2){$[_0x1020('‫b2','SWuO')](_0x29b2f2,_0xca23c6);}finally{_0x2f0332[_0x1020('‫b3','Daw0')](_0x28369);}}else{return'1';}});});}function getRandomIDPro(){var _0x4a1c80={'MCEiN':function(_0x16b8fa,_0x2d0b9e){return _0x16b8fa===_0x2d0b9e;},'jahUQ':function(_0x2d4546,_0xd8a84b){return _0x2d4546<_0xd8a84b;},'iSftr':function(_0x3f766f,_0x5e7e2e){return _0x3f766f!==_0x5e7e2e;},'wBBSl':_0x1020('‫b4','dNjI'),'zxzfA':function(_0x3357f7,_0xfa7f26){return _0x3357f7==_0xfa7f26;},'RtARN':_0x1020('‫b5','UO9)'),'hWrqL':_0x1020('‫b6','b2cz'),'FmeaU':_0x1020('‮b7','uAZE'),'ZRHUz':_0x1020('‫b8','[NNj'),'bvXvE':_0x1020('‮b9','UO9)'),'wmfAL':_0x1020('‮ba','b2cz'),'QAdHS':_0x1020('‫bb','UO9)'),'ODQPn':function(_0x45d47e,_0x18a571){return _0x45d47e|_0x18a571;},'DxfAX':function(_0x2bed6e,_0x1c2fbd){return _0x2bed6e*_0x1c2fbd;},'OvlFu':function(_0x49bd15,_0x4997ad){return _0x49bd15>_0x4997ad;}};var _0x318a4c,_0x35d396,_0x2c68b9=_0x4a1c80[_0x1020('‮bc','MI]9')](void 0x0,_0x44fbbc=(_0x35d396=_0x4a1c80[_0x1020('‫bd',']*hE')](0x0,arguments[_0x1020('‫be',')nkr')])&&_0x4a1c80[_0x1020('‮bf','b2cz')](void 0x0,arguments[0x0])?arguments[0x0]:{})[_0x1020('‫c0','WtE@')])?0xa:_0x44fbbc,_0x44fbbc=_0x4a1c80[_0x1020('‫c1','$l&B')](void 0x0,_0x44fbbc=_0x35d396[_0x1020('‮c2','v1*y')])?_0x4a1c80[_0x1020('‫c3','sQfA')]:_0x44fbbc,_0x3eb851='';if((_0x35d396=_0x35d396[_0x1020('‮c4',')jdl')])&&_0x4a1c80[_0x1020('‮c5','b2cz')](_0x4a1c80[_0x1020('‮c6','VL82')],typeof _0x35d396))_0x318a4c=_0x35d396;else switch(_0x44fbbc){case _0x4a1c80[_0x1020('‫c7','@1ER')]:_0x318a4c=_0x4a1c80[_0x1020('‫c8','1^Ap')];break;case _0x4a1c80[_0x1020('‮c9',']*hE')]:_0x318a4c=_0x4a1c80[_0x1020('‫ca','oqeR')];break;case _0x4a1c80[_0x1020('‫cb','S5sc')]:default:_0x318a4c=_0x4a1c80[_0x1020('‫cc','j8Xt')];}if(_0x4a1c80[_0x1020('‫cd','zi7G')]($[_0x1020('‮ce','goYs')][_0x1020('‮cf','@1ER')](_0x4a1c80[_0x1020('‫d0','C0[W')]),-0x1))return'1';for(;_0x2c68b9--;)_0x3eb851+=_0x318a4c[_0x4a1c80[_0x1020('‫d1','WD@s')](_0x4a1c80[_0x1020('‫d2','dNjI')](Math[_0x1020('‮d3','wCUy')](),_0x318a4c[_0x1020('‫be',')nkr')]),0x0)];if(_0x4a1c80[_0x1020('‫d4','DI8o')](new Date()[_0x1020('‮d5','QtPE')](),0x18016fe3480)){return'1';}return _0x3eb851;};_0xodh='jsjiami.com.v6';
 
 function CryptoScripts() {
   // prettier-ignore
   !function(t,e){"object"==typeof exports?module.exports=exports=e():"function"==typeof define&&define.amd?define([],e):t.CryptoJS=e()}(this,function(){var t,e,r,i,n,o,s,c,a,h,l,f,d,u,p,_,v,y,g,B,w,k,S,m,x,b,H,z,A,C,D,E,R,M,F,P,W,O,I,U,K,X,L,j,N,T,q,Z,V,G,J,$,Q,Y,tt,et,rt,it,nt,ot,st,ct,at,ht,lt,ft,dt,ut,pt,_t,vt,yt,gt,Bt,wt,kt,St,mt=mt||function(t){var e;if("undefined"!=typeof window&&window.crypto&&(e=window.crypto),!e&&"undefined"!=typeof window&&window.msCrypto&&(e=window.msCrypto),!e&&"undefined"!=typeof global&&global.crypto&&(e=global.crypto),!e&&"function"==typeof require)try{e=require("crypto")}catch(e){}function r(){if(e){if("function"==typeof e.getRandomValues)try{return e.getRandomValues(new Uint32Array(1))[0]}catch(t){}if("function"==typeof e.randomBytes)try{return e.randomBytes(4).readInt32LE()}catch(t){}}throw new Error("Native crypto module could not be used to get secure random number.")}var i=Object.create||function(t){var e;return n.prototype=t,e=new n,n.prototype=null,e};function n(){}var o={},s=o.lib={},c=s.Base={extend:function(t){var e=i(this);return t&&e.mixIn(t),e.hasOwnProperty("init")&&this.init!==e.init||(e.init=function(){e.$super.init.apply(this,arguments)}),(e.init.prototype=e).$super=this,e},create:function(){var t=this.extend();return t.init.apply(t,arguments),t},init:function(){},mixIn:function(t){for(var e in t)t.hasOwnProperty(e)&&(this[e]=t[e]);t.hasOwnProperty("toString")&&(this.toString=t.toString)},clone:function(){return this.init.prototype.extend(this)}},a=s.WordArray=c.extend({init:function(t,e){t=this.words=t||[],this.sigBytes=null!=e?e:4*t.length},toString:function(t){return(t||l).stringify(this)},concat:function(t){var e=this.words,r=t.words,i=this.sigBytes,n=t.sigBytes;if(this.clamp(),i%4)for(var o=0;o<n;o++){var s=r[o>>>2]>>>24-o%4*8&255;e[i+o>>>2]|=s<<24-(i+o)%4*8}else for(o=0;o<n;o+=4)e[i+o>>>2]=r[o>>>2];return this.sigBytes+=n,this},clamp:function(){var e=this.words,r=this.sigBytes;e[r>>>2]&=4294967295<<32-r%4*8,e.length=t.ceil(r/4)},clone:function(){var t=c.clone.call(this);return t.words=this.words.slice(0),t},random:function(t){for(var e=[],i=0;i<t;i+=4)e.push(r());return new a.init(e,t)}}),h=o.enc={},l=h.Hex={stringify:function(t){for(var e=t.words,r=t.sigBytes,i=[],n=0;n<r;n++){var o=e[n>>>2]>>>24-n%4*8&255;i.push((o>>>4).toString(16)),i.push((15&o).toString(16))}return i.join("")},parse:function(t){for(var e=t.length,r=[],i=0;i<e;i+=2)r[i>>>3]|=parseInt(t.substr(i,2),16)<<24-i%8*4;return new a.init(r,e/2)}},f=h.Latin1={stringify:function(t){for(var e=t.words,r=t.sigBytes,i=[],n=0;n<r;n++){var o=e[n>>>2]>>>24-n%4*8&255;i.push(String.fromCharCode(o))}return i.join("")},parse:function(t){for(var e=t.length,r=[],i=0;i<e;i++)r[i>>>2]|=(255&t.charCodeAt(i))<<24-i%4*8;return new a.init(r,e)}},d=h.Utf8={stringify:function(t){try{return decodeURIComponent(escape(f.stringify(t)))}catch(t){throw new Error("Malformed UTF-8 data")}},parse:function(t){return f.parse(unescape(encodeURIComponent(t)))}},u=s.BufferedBlockAlgorithm=c.extend({reset:function(){this._data=new a.init,this._nDataBytes=0},_append:function(t){"string"==typeof t&&(t=d.parse(t)),this._data.concat(t),this._nDataBytes+=t.sigBytes},_process:function(e){var r,i=this._data,n=i.words,o=i.sigBytes,s=this.blockSize,c=o/(4*s),h=(c=e?t.ceil(c):t.max((0|c)-this._minBufferSize,0))*s,l=t.min(4*h,o);if(h){for(var f=0;f<h;f+=s)this._doProcessBlock(n,f);r=n.splice(0,h),i.sigBytes-=l}return new a.init(r,l)},clone:function(){var t=c.clone.call(this);return t._data=this._data.clone(),t},_minBufferSize:0}),p=(s.Hasher=u.extend({cfg:c.extend(),init:function(t){this.cfg=this.cfg.extend(t),this.reset()},reset:function(){u.reset.call(this),this._doReset()},update:function(t){return this._append(t),this._process(),this},finalize:function(t){return t&&this._append(t),this._doFinalize()},blockSize:16,_createHelper:function(t){return function(e,r){return new t.init(r).finalize(e)}},_createHmacHelper:function(t){return function(e,r){return new p.HMAC.init(t,r).finalize(e)}}}),o.algo={});return o}(Math);function xt(t,e,r){return t^e^r}function bt(t,e,r){return t&e|~t&r}function Ht(t,e,r){return(t|~e)^r}function zt(t,e,r){return t&r|e&~r}function At(t,e,r){return t^(e|~r)}function Ct(t,e){return t<<e|t>>>32-e}function Dt(t,e,r,i){var n,o=this._iv;o?(n=o.slice(0),this._iv=void 0):n=this._prevBlock,i.encryptBlock(n,0);for(var s=0;s<r;s++)t[e+s]^=n[s]}function Et(t){if(255==(t>>24&255)){var e=t>>16&255,r=t>>8&255,i=255&t;255===e?(e=0,255===r?(r=0,255===i?i=0:++i):++r):++e,t=0,t+=e<<16,t+=r<<8,t+=i}else t+=1<<24;return t}function Rt(){for(var t=this._X,e=this._C,r=0;r<8;r++)ft[r]=e[r];for(e[0]=e[0]+1295307597+this._b|0,e[1]=e[1]+3545052371+(e[0]>>>0<ft[0]>>>0?1:0)|0,e[2]=e[2]+886263092+(e[1]>>>0<ft[1]>>>0?1:0)|0,e[3]=e[3]+1295307597+(e[2]>>>0<ft[2]>>>0?1:0)|0,e[4]=e[4]+3545052371+(e[3]>>>0<ft[3]>>>0?1:0)|0,e[5]=e[5]+886263092+(e[4]>>>0<ft[4]>>>0?1:0)|0,e[6]=e[6]+1295307597+(e[5]>>>0<ft[5]>>>0?1:0)|0,e[7]=e[7]+3545052371+(e[6]>>>0<ft[6]>>>0?1:0)|0,this._b=e[7]>>>0<ft[7]>>>0?1:0,r=0;r<8;r++){var i=t[r]+e[r],n=65535&i,o=i>>>16,s=((n*n>>>17)+n*o>>>15)+o*o,c=((4294901760&i)*i|0)+((65535&i)*i|0);dt[r]=s^c}t[0]=dt[0]+(dt[7]<<16|dt[7]>>>16)+(dt[6]<<16|dt[6]>>>16)|0,t[1]=dt[1]+(dt[0]<<8|dt[0]>>>24)+dt[7]|0,t[2]=dt[2]+(dt[1]<<16|dt[1]>>>16)+(dt[0]<<16|dt[0]>>>16)|0,t[3]=dt[3]+(dt[2]<<8|dt[2]>>>24)+dt[1]|0,t[4]=dt[4]+(dt[3]<<16|dt[3]>>>16)+(dt[2]<<16|dt[2]>>>16)|0,t[5]=dt[5]+(dt[4]<<8|dt[4]>>>24)+dt[3]|0,t[6]=dt[6]+(dt[5]<<16|dt[5]>>>16)+(dt[4]<<16|dt[4]>>>16)|0,t[7]=dt[7]+(dt[6]<<8|dt[6]>>>24)+dt[5]|0}function Mt(){for(var t=this._X,e=this._C,r=0;r<8;r++)wt[r]=e[r];for(e[0]=e[0]+1295307597+this._b|0,e[1]=e[1]+3545052371+(e[0]>>>0<wt[0]>>>0?1:0)|0,e[2]=e[2]+886263092+(e[1]>>>0<wt[1]>>>0?1:0)|0,e[3]=e[3]+1295307597+(e[2]>>>0<wt[2]>>>0?1:0)|0,e[4]=e[4]+3545052371+(e[3]>>>0<wt[3]>>>0?1:0)|0,e[5]=e[5]+886263092+(e[4]>>>0<wt[4]>>>0?1:0)|0,e[6]=e[6]+1295307597+(e[5]>>>0<wt[5]>>>0?1:0)|0,e[7]=e[7]+3545052371+(e[6]>>>0<wt[6]>>>0?1:0)|0,this._b=e[7]>>>0<wt[7]>>>0?1:0,r=0;r<8;r++){var i=t[r]+e[r],n=65535&i,o=i>>>16,s=((n*n>>>17)+n*o>>>15)+o*o,c=((4294901760&i)*i|0)+((65535&i)*i|0);kt[r]=s^c}t[0]=kt[0]+(kt[7]<<16|kt[7]>>>16)+(kt[6]<<16|kt[6]>>>16)|0,t[1]=kt[1]+(kt[0]<<8|kt[0]>>>24)+kt[7]|0,t[2]=kt[2]+(kt[1]<<16|kt[1]>>>16)+(kt[0]<<16|kt[0]>>>16)|0,t[3]=kt[3]+(kt[2]<<8|kt[2]>>>24)+kt[1]|0,t[4]=kt[4]+(kt[3]<<16|kt[3]>>>16)+(kt[2]<<16|kt[2]>>>16)|0,t[5]=kt[5]+(kt[4]<<8|kt[4]>>>24)+kt[3]|0,t[6]=kt[6]+(kt[5]<<16|kt[5]>>>16)+(kt[4]<<16|kt[4]>>>16)|0,t[7]=kt[7]+(kt[6]<<8|kt[6]>>>24)+kt[5]|0}return t=mt.lib.WordArray,mt.enc.Base64={stringify:function(t){var e=t.words,r=t.sigBytes,i=this._map;t.clamp();for(var n=[],o=0;o<r;o+=3)for(var s=(e[o>>>2]>>>24-o%4*8&255)<<16|(e[o+1>>>2]>>>24-(o+1)%4*8&255)<<8|e[o+2>>>2]>>>24-(o+2)%4*8&255,c=0;c<4&&o+.75*c<r;c++)n.push(i.charAt(s>>>6*(3-c)&63));var a=i.charAt(64);if(a)for(;n.length%4;)n.push(a);return n.join("")},parse:function(e){var r=e.length,i=this._map,n=this._reverseMap;if(!n){n=this._reverseMap=[];for(var o=0;o<i.length;o++)n[i.charCodeAt(o)]=o}var s=i.charAt(64);if(s){var c=e.indexOf(s);-1!==c&&(r=c)}return function(e,r,i){for(var n=[],o=0,s=0;s<r;s++)if(s%4){var c=i[e.charCodeAt(s-1)]<<s%4*2|i[e.charCodeAt(s)]>>>6-s%4*2;n[o>>>2]|=c<<24-o%4*8,o++}return t.create(n,o)}(e,r,n)},_map:"ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/="},function(t){var e=mt,r=e.lib,i=r.WordArray,n=r.Hasher,o=e.algo,s=[];!function(){for(var e=0;e<64;e++)s[e]=4294967296*t.abs(t.sin(e+1))|0}();var c=o.MD5=n.extend({_doReset:function(){this._hash=new i.init([1732584193,4023233417,2562383102,271733878])},_doProcessBlock:function(t,e){for(var r=0;r<16;r++){var i=e+r,n=t[i];t[i]=16711935&(n<<8|n>>>24)|4278255360&(n<<24|n>>>8)}var o=this._hash.words,c=t[e+0],d=t[e+1],u=t[e+2],p=t[e+3],_=t[e+4],v=t[e+5],y=t[e+6],g=t[e+7],B=t[e+8],w=t[e+9],k=t[e+10],S=t[e+11],m=t[e+12],x=t[e+13],b=t[e+14],H=t[e+15],z=o[0],A=o[1],C=o[2],D=o[3];z=f(z=l(z=l(z=l(z=l(z=h(z=h(z=h(z=h(z=a(z=a(z=a(z=a(z,A,C,D,c,7,s[0]),A=a(A,C=a(C,D=a(D,z,A,C,d,12,s[1]),z,A,u,17,s[2]),D,z,p,22,s[3]),C,D,_,7,s[4]),A=a(A,C=a(C,D=a(D,z,A,C,v,12,s[5]),z,A,y,17,s[6]),D,z,g,22,s[7]),C,D,B,7,s[8]),A=a(A,C=a(C,D=a(D,z,A,C,w,12,s[9]),z,A,k,17,s[10]),D,z,S,22,s[11]),C,D,m,7,s[12]),A=a(A,C=a(C,D=a(D,z,A,C,x,12,s[13]),z,A,b,17,s[14]),D,z,H,22,s[15]),C,D,d,5,s[16]),A=h(A,C=h(C,D=h(D,z,A,C,y,9,s[17]),z,A,S,14,s[18]),D,z,c,20,s[19]),C,D,v,5,s[20]),A=h(A,C=h(C,D=h(D,z,A,C,k,9,s[21]),z,A,H,14,s[22]),D,z,_,20,s[23]),C,D,w,5,s[24]),A=h(A,C=h(C,D=h(D,z,A,C,b,9,s[25]),z,A,p,14,s[26]),D,z,B,20,s[27]),C,D,x,5,s[28]),A=h(A,C=h(C,D=h(D,z,A,C,u,9,s[29]),z,A,g,14,s[30]),D,z,m,20,s[31]),C,D,v,4,s[32]),A=l(A,C=l(C,D=l(D,z,A,C,B,11,s[33]),z,A,S,16,s[34]),D,z,b,23,s[35]),C,D,d,4,s[36]),A=l(A,C=l(C,D=l(D,z,A,C,_,11,s[37]),z,A,g,16,s[38]),D,z,k,23,s[39]),C,D,x,4,s[40]),A=l(A,C=l(C,D=l(D,z,A,C,c,11,s[41]),z,A,p,16,s[42]),D,z,y,23,s[43]),C,D,w,4,s[44]),A=l(A,C=l(C,D=l(D,z,A,C,m,11,s[45]),z,A,H,16,s[46]),D,z,u,23,s[47]),C,D,c,6,s[48]),A=f(A=f(A=f(A=f(A,C=f(C,D=f(D,z,A,C,g,10,s[49]),z,A,b,15,s[50]),D,z,v,21,s[51]),C=f(C,D=f(D,z=f(z,A,C,D,m,6,s[52]),A,C,p,10,s[53]),z,A,k,15,s[54]),D,z,d,21,s[55]),C=f(C,D=f(D,z=f(z,A,C,D,B,6,s[56]),A,C,H,10,s[57]),z,A,y,15,s[58]),D,z,x,21,s[59]),C=f(C,D=f(D,z=f(z,A,C,D,_,6,s[60]),A,C,S,10,s[61]),z,A,u,15,s[62]),D,z,w,21,s[63]),o[0]=o[0]+z|0,o[1]=o[1]+A|0,o[2]=o[2]+C|0,o[3]=o[3]+D|0},_doFinalize:function(){var e=this._data,r=e.words,i=8*this._nDataBytes,n=8*e.sigBytes;r[n>>>5]|=128<<24-n%32;var o=t.floor(i/4294967296),s=i;r[15+(64+n>>>9<<4)]=16711935&(o<<8|o>>>24)|4278255360&(o<<24|o>>>8),r[14+(64+n>>>9<<4)]=16711935&(s<<8|s>>>24)|4278255360&(s<<24|s>>>8),e.sigBytes=4*(r.length+1),this._process();for(var c=this._hash,a=c.words,h=0;h<4;h++){var l=a[h];a[h]=16711935&(l<<8|l>>>24)|4278255360&(l<<24|l>>>8)}return c},clone:function(){var t=n.clone.call(this);return t._hash=this._hash.clone(),t}});function a(t,e,r,i,n,o,s){var c=t+(e&r|~e&i)+n+s;return(c<<o|c>>>32-o)+e}function h(t,e,r,i,n,o,s){var c=t+(e&i|r&~i)+n+s;return(c<<o|c>>>32-o)+e}function l(t,e,r,i,n,o,s){var c=t+(e^r^i)+n+s;return(c<<o|c>>>32-o)+e}function f(t,e,r,i,n,o,s){var c=t+(r^(e|~i))+n+s;return(c<<o|c>>>32-o)+e}e.MD5=n._createHelper(c),e.HmacMD5=n._createHmacHelper(c)}(Math),r=(e=mt).lib,i=r.WordArray,n=r.Hasher,o=e.algo,s=[],c=o.SHA1=n.extend({_doReset:function(){this._hash=new i.init([1732584193,4023233417,2562383102,271733878,3285377520])},_doProcessBlock:function(t,e){for(var r=this._hash.words,i=r[0],n=r[1],o=r[2],c=r[3],a=r[4],h=0;h<80;h++){if(h<16)s[h]=0|t[e+h];else{var l=s[h-3]^s[h-8]^s[h-14]^s[h-16];s[h]=l<<1|l>>>31}var f=(i<<5|i>>>27)+a+s[h];f+=h<20?1518500249+(n&o|~n&c):h<40?1859775393+(n^o^c):h<60?(n&o|n&c|o&c)-1894007588:(n^o^c)-899497514,a=c,c=o,o=n<<30|n>>>2,n=i,i=f}r[0]=r[0]+i|0,r[1]=r[1]+n|0,r[2]=r[2]+o|0,r[3]=r[3]+c|0,r[4]=r[4]+a|0},_doFinalize:function(){var t=this._data,e=t.words,r=8*this._nDataBytes,i=8*t.sigBytes;return e[i>>>5]|=128<<24-i%32,e[14+(64+i>>>9<<4)]=Math.floor(r/4294967296),e[15+(64+i>>>9<<4)]=r,t.sigBytes=4*e.length,this._process(),this._hash},clone:function(){var t=n.clone.call(this);return t._hash=this._hash.clone(),t}}),e.SHA1=n._createHelper(c),e.HmacSHA1=n._createHmacHelper(c),function(t){var e=mt,r=e.lib,i=r.WordArray,n=r.Hasher,o=e.algo,s=[],c=[];!function(){function e(e){for(var r=t.sqrt(e),i=2;i<=r;i++)if(!(e%i))return;return 1}function r(t){return 4294967296*(t-(0|t))|0}for(var i=2,n=0;n<64;)e(i)&&(n<8&&(s[n]=r(t.pow(i,.5))),c[n]=r(t.pow(i,1/3)),n++),i++}();var a=[],h=o.SHA256=n.extend({_doReset:function(){this._hash=new i.init(s.slice(0))},_doProcessBlock:function(t,e){for(var r=this._hash.words,i=r[0],n=r[1],o=r[2],s=r[3],h=r[4],l=r[5],f=r[6],d=r[7],u=0;u<64;u++){if(u<16)a[u]=0|t[e+u];else{var p=a[u-15],_=(p<<25|p>>>7)^(p<<14|p>>>18)^p>>>3,v=a[u-2],y=(v<<15|v>>>17)^(v<<13|v>>>19)^v>>>10;a[u]=_+a[u-7]+y+a[u-16]}var g=i&n^i&o^n&o,B=(i<<30|i>>>2)^(i<<19|i>>>13)^(i<<10|i>>>22),w=d+((h<<26|h>>>6)^(h<<21|h>>>11)^(h<<7|h>>>25))+(h&l^~h&f)+c[u]+a[u];d=f,f=l,l=h,h=s+w|0,s=o,o=n,n=i,i=w+(B+g)|0}r[0]=r[0]+i|0,r[1]=r[1]+n|0,r[2]=r[2]+o|0,r[3]=r[3]+s|0,r[4]=r[4]+h|0,r[5]=r[5]+l|0,r[6]=r[6]+f|0,r[7]=r[7]+d|0},_doFinalize:function(){var e=this._data,r=e.words,i=8*this._nDataBytes,n=8*e.sigBytes;return r[n>>>5]|=128<<24-n%32,r[14+(64+n>>>9<<4)]=t.floor(i/4294967296),r[15+(64+n>>>9<<4)]=i,e.sigBytes=4*r.length,this._process(),this._hash},clone:function(){var t=n.clone.call(this);return t._hash=this._hash.clone(),t}});e.SHA256=n._createHelper(h),e.HmacSHA256=n._createHmacHelper(h)}(Math),function(){var t=mt.lib.WordArray,e=mt.enc;function r(t){return t<<8&4278255360|t>>>8&16711935}e.Utf16=e.Utf16BE={stringify:function(t){for(var e=t.words,r=t.sigBytes,i=[],n=0;n<r;n+=2){var o=e[n>>>2]>>>16-n%4*8&65535;i.push(String.fromCharCode(o))}return i.join("")},parse:function(e){for(var r=e.length,i=[],n=0;n<r;n++)i[n>>>1]|=e.charCodeAt(n)<<16-n%2*16;return t.create(i,2*r)}},e.Utf16LE={stringify:function(t){for(var e=t.words,i=t.sigBytes,n=[],o=0;o<i;o+=2){var s=r(e[o>>>2]>>>16-o%4*8&65535);n.push(String.fromCharCode(s))}return n.join("")},parse:function(e){for(var i=e.length,n=[],o=0;o<i;o++)n[o>>>1]|=r(e.charCodeAt(o)<<16-o%2*16);return t.create(n,2*i)}}}(),function(){if("function"==typeof ArrayBuffer){var t=mt.lib.WordArray,e=t.init;(t.init=function(t){if(t instanceof ArrayBuffer&&(t=new Uint8Array(t)),(t instanceof Int8Array||"undefined"!=typeof Uint8ClampedArray&&t instanceof Uint8ClampedArray||t instanceof Int16Array||t instanceof Uint16Array||t instanceof Int32Array||t instanceof Uint32Array||t instanceof Float32Array||t instanceof Float64Array)&&(t=new Uint8Array(t.buffer,t.byteOffset,t.byteLength)),t instanceof Uint8Array){for(var r=t.byteLength,i=[],n=0;n<r;n++)i[n>>>2]|=t[n]<<24-n%4*8;e.call(this,i,r)}else e.apply(this,arguments)}).prototype=t}}(),Math,h=(a=mt).lib,l=h.WordArray,f=h.Hasher,d=a.algo,u=l.create([0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,7,4,13,1,10,6,15,3,12,0,9,5,2,14,11,8,3,10,14,4,9,15,8,1,2,7,0,6,13,11,5,12,1,9,11,10,0,8,12,4,13,3,7,15,14,5,6,2,4,0,5,9,7,12,2,10,14,1,3,8,11,6,15,13]),p=l.create([5,14,7,0,9,2,11,4,13,6,15,8,1,10,3,12,6,11,3,7,0,13,5,10,14,15,8,12,4,9,1,2,15,5,1,3,7,14,6,9,11,8,12,2,10,0,4,13,8,6,4,1,3,11,15,0,5,12,2,13,9,7,10,14,12,15,10,4,1,5,8,7,6,2,13,14,0,3,9,11]),_=l.create([11,14,15,12,5,8,7,9,11,13,14,15,6,7,9,8,7,6,8,13,11,9,7,15,7,12,15,9,11,7,13,12,11,13,6,7,14,9,13,15,14,8,13,6,5,12,7,5,11,12,14,15,14,15,9,8,9,14,5,6,8,6,5,12,9,15,5,11,6,8,13,12,5,12,13,14,11,8,5,6]),v=l.create([8,9,9,11,13,15,15,5,7,7,8,11,14,14,12,6,9,13,15,7,12,8,9,11,7,7,12,7,6,15,13,11,9,7,15,11,8,6,6,14,12,13,5,14,13,13,7,5,15,5,8,11,14,14,6,14,6,9,12,9,12,5,15,8,8,5,12,9,12,5,14,6,8,13,6,5,15,13,11,11]),y=l.create([0,1518500249,1859775393,2400959708,2840853838]),g=l.create([1352829926,1548603684,1836072691,2053994217,0]),B=d.RIPEMD160=f.extend({_doReset:function(){this._hash=l.create([1732584193,4023233417,2562383102,271733878,3285377520])},_doProcessBlock:function(t,e){for(var r=0;r<16;r++){var i=e+r,n=t[i];t[i]=16711935&(n<<8|n>>>24)|4278255360&(n<<24|n>>>8)}var o,s,c,a,h,l,f,d,B,w,k,S=this._hash.words,m=y.words,x=g.words,b=u.words,H=p.words,z=_.words,A=v.words;for(l=o=S[0],f=s=S[1],d=c=S[2],B=a=S[3],w=h=S[4],r=0;r<80;r+=1)k=o+t[e+b[r]]|0,k+=r<16?xt(s,c,a)+m[0]:r<32?bt(s,c,a)+m[1]:r<48?Ht(s,c,a)+m[2]:r<64?zt(s,c,a)+m[3]:At(s,c,a)+m[4],k=(k=Ct(k|=0,z[r]))+h|0,o=h,h=a,a=Ct(c,10),c=s,s=k,k=l+t[e+H[r]]|0,k+=r<16?At(f,d,B)+x[0]:r<32?zt(f,d,B)+x[1]:r<48?Ht(f,d,B)+x[2]:r<64?bt(f,d,B)+x[3]:xt(f,d,B)+x[4],k=(k=Ct(k|=0,A[r]))+w|0,l=w,w=B,B=Ct(d,10),d=f,f=k;k=S[1]+c+B|0,S[1]=S[2]+a+w|0,S[2]=S[3]+h+l|0,S[3]=S[4]+o+f|0,S[4]=S[0]+s+d|0,S[0]=k},_doFinalize:function(){var t=this._data,e=t.words,r=8*this._nDataBytes,i=8*t.sigBytes;e[i>>>5]|=128<<24-i%32,e[14+(64+i>>>9<<4)]=16711935&(r<<8|r>>>24)|4278255360&(r<<24|r>>>8),t.sigBytes=4*(e.length+1),this._process();for(var n=this._hash,o=n.words,s=0;s<5;s++){var c=o[s];o[s]=16711935&(c<<8|c>>>24)|4278255360&(c<<24|c>>>8)}return n},clone:function(){var t=f.clone.call(this);return t._hash=this._hash.clone(),t}}),a.RIPEMD160=f._createHelper(B),a.HmacRIPEMD160=f._createHmacHelper(B),w=mt.lib.Base,k=mt.enc.Utf8,mt.algo.HMAC=w.extend({init:function(t,e){t=this._hasher=new t.init,"string"==typeof e&&(e=k.parse(e));var r=t.blockSize,i=4*r;e.sigBytes>i&&(e=t.finalize(e)),e.clamp();for(var n=this._oKey=e.clone(),o=this._iKey=e.clone(),s=n.words,c=o.words,a=0;a<r;a++)s[a]^=1549556828,c[a]^=909522486;n.sigBytes=o.sigBytes=i,this.reset()},reset:function(){var t=this._hasher;t.reset(),t.update(this._iKey)},update:function(t){return this._hasher.update(t),this},finalize:function(t){var e=this._hasher,r=e.finalize(t);return e.reset(),e.finalize(this._oKey.clone().concat(r))}}),x=(m=(S=mt).lib).Base,b=m.WordArray,z=(H=S.algo).SHA1,A=H.HMAC,C=H.PBKDF2=x.extend({cfg:x.extend({keySize:4,hasher:z,iterations:1}),init:function(t){this.cfg=this.cfg.extend(t)},compute:function(t,e){for(var r=this.cfg,i=A.create(r.hasher,t),n=b.create(),o=b.create([1]),s=n.words,c=o.words,a=r.keySize,h=r.iterations;s.length<a;){var l=i.update(e).finalize(o);i.reset();for(var f=l.words,d=f.length,u=l,p=1;p<h;p++){u=i.finalize(u),i.reset();for(var _=u.words,v=0;v<d;v++)f[v]^=_[v]}n.concat(l),c[0]++}return n.sigBytes=4*a,n}}),S.PBKDF2=function(t,e,r){return C.create(r).compute(t,e)},R=(E=(D=mt).lib).Base,M=E.WordArray,P=(F=D.algo).MD5,W=F.EvpKDF=R.extend({cfg:R.extend({keySize:4,hasher:P,iterations:1}),init:function(t){this.cfg=this.cfg.extend(t)},compute:function(t,e){for(var r,i=this.cfg,n=i.hasher.create(),o=M.create(),s=o.words,c=i.keySize,a=i.iterations;s.length<c;){r&&n.update(r),r=n.update(t).finalize(e),n.reset();for(var h=1;h<a;h++)r=n.finalize(r),n.reset();o.concat(r)}return o.sigBytes=4*c,o}}),D.EvpKDF=function(t,e,r){return W.create(r).compute(t,e)},I=(O=mt).lib.WordArray,U=O.algo,K=U.SHA256,X=U.SHA224=K.extend({_doReset:function(){this._hash=new I.init([3238371032,914150663,812702999,4144912697,4290775857,1750603025,1694076839,3204075428])},_doFinalize:function(){var t=K._doFinalize.call(this);return t.sigBytes-=4,t}}),O.SHA224=K._createHelper(X),O.HmacSHA224=K._createHmacHelper(X),L=mt.lib,j=L.Base,N=L.WordArray,(T=mt.x64={}).Word=j.extend({init:function(t,e){this.high=t,this.low=e}}),T.WordArray=j.extend({init:function(t,e){t=this.words=t||[],this.sigBytes=null!=e?e:8*t.length},toX32:function(){for(var t=this.words,e=t.length,r=[],i=0;i<e;i++){var n=t[i];r.push(n.high),r.push(n.low)}return N.create(r,this.sigBytes)},clone:function(){for(var t=j.clone.call(this),e=t.words=this.words.slice(0),r=e.length,i=0;i<r;i++)e[i]=e[i].clone();return t}}),function(t){var e=mt,r=e.lib,i=r.WordArray,n=r.Hasher,o=e.x64.Word,s=e.algo,c=[],a=[],h=[];!function(){for(var t=1,e=0,r=0;r<24;r++){c[t+5*e]=(r+1)*(r+2)/2%64;var i=(2*t+3*e)%5;t=e%5,e=i}for(t=0;t<5;t++)for(e=0;e<5;e++)a[t+5*e]=e+(2*t+3*e)%5*5;for(var n=1,s=0;s<24;s++){for(var l=0,f=0,d=0;d<7;d++){if(1&n){var u=(1<<d)-1;u<32?f^=1<<u:l^=1<<u-32}128&n?n=n<<1^113:n<<=1}h[s]=o.create(l,f)}}();var l=[];!function(){for(var t=0;t<25;t++)l[t]=o.create()}();var f=s.SHA3=n.extend({cfg:n.cfg.extend({outputLength:512}),_doReset:function(){for(var t=this._state=[],e=0;e<25;e++)t[e]=new o.init;this.blockSize=(1600-2*this.cfg.outputLength)/32},_doProcessBlock:function(t,e){for(var r=this._state,i=this.blockSize/2,n=0;n<i;n++){var o=t[e+2*n],s=t[e+2*n+1];o=16711935&(o<<8|o>>>24)|4278255360&(o<<24|o>>>8),s=16711935&(s<<8|s>>>24)|4278255360&(s<<24|s>>>8),(A=r[n]).high^=s,A.low^=o}for(var f=0;f<24;f++){for(var d=0;d<5;d++){for(var u=0,p=0,_=0;_<5;_++)u^=(A=r[d+5*_]).high,p^=A.low;var v=l[d];v.high=u,v.low=p}for(d=0;d<5;d++){var y=l[(d+4)%5],g=l[(d+1)%5],B=g.high,w=g.low;for(u=y.high^(B<<1|w>>>31),p=y.low^(w<<1|B>>>31),_=0;_<5;_++)(A=r[d+5*_]).high^=u,A.low^=p}for(var k=1;k<25;k++){var S=(A=r[k]).high,m=A.low,x=c[k];p=x<32?(u=S<<x|m>>>32-x,m<<x|S>>>32-x):(u=m<<x-32|S>>>64-x,S<<x-32|m>>>64-x);var b=l[a[k]];b.high=u,b.low=p}var H=l[0],z=r[0];for(H.high=z.high,H.low=z.low,d=0;d<5;d++)for(_=0;_<5;_++){var A=r[k=d+5*_],C=l[k],D=l[(d+1)%5+5*_],E=l[(d+2)%5+5*_];A.high=C.high^~D.high&E.high,A.low=C.low^~D.low&E.low}A=r[0];var R=h[f];A.high^=R.high,A.low^=R.low}},_doFinalize:function(){var e=this._data,r=e.words,n=(this._nDataBytes,8*e.sigBytes),o=32*this.blockSize;r[n>>>5]|=1<<24-n%32,r[(t.ceil((1+n)/o)*o>>>5)-1]|=128,e.sigBytes=4*r.length,this._process();for(var s=this._state,c=this.cfg.outputLength/8,a=c/8,h=[],l=0;l<a;l++){var f=s[l],d=f.high,u=f.low;d=16711935&(d<<8|d>>>24)|4278255360&(d<<24|d>>>8),u=16711935&(u<<8|u>>>24)|4278255360&(u<<24|u>>>8),h.push(u),h.push(d)}return new i.init(h,c)},clone:function(){for(var t=n.clone.call(this),e=t._state=this._state.slice(0),r=0;r<25;r++)e[r]=e[r].clone();return t}});e.SHA3=n._createHelper(f),e.HmacSHA3=n._createHmacHelper(f)}(Math),function(){var t=mt,e=t.lib.Hasher,r=t.x64,i=r.Word,n=r.WordArray,o=t.algo;function s(){return i.create.apply(i,arguments)}var c=[s(1116352408,3609767458),s(1899447441,602891725),s(3049323471,3964484399),s(3921009573,2173295548),s(961987163,4081628472),s(1508970993,3053834265),s(2453635748,2937671579),s(2870763221,3664609560),s(3624381080,2734883394),s(310598401,1164996542),s(607225278,1323610764),s(1426881987,3590304994),s(1925078388,4068182383),s(2162078206,991336113),s(2614888103,633803317),s(3248222580,3479774868),s(3835390401,2666613458),s(4022224774,944711139),s(264347078,2341262773),s(604807628,2007800933),s(770255983,1495990901),s(1249150122,1856431235),s(1555081692,3175218132),s(1996064986,2198950837),s(2554220882,3999719339),s(2821834349,766784016),s(2952996808,2566594879),s(3210313671,3203337956),s(3336571891,1034457026),s(3584528711,2466948901),s(113926993,3758326383),s(338241895,168717936),s(666307205,1188179964),s(773529912,1546045734),s(1294757372,1522805485),s(1396182291,2643833823),s(1695183700,2343527390),s(1986661051,1014477480),s(2177026350,1206759142),s(2456956037,344077627),s(2730485921,1290863460),s(2820302411,3158454273),s(3259730800,3505952657),s(3345764771,106217008),s(3516065817,3606008344),s(3600352804,1432725776),s(4094571909,1467031594),s(275423344,851169720),s(430227734,3100823752),s(506948616,1363258195),s(659060556,3750685593),s(883997877,3785050280),s(958139571,3318307427),s(1322822218,3812723403),s(1537002063,2003034995),s(1747873779,3602036899),s(1955562222,1575990012),s(2024104815,1125592928),s(2227730452,2716904306),s(2361852424,442776044),s(2428436474,593698344),s(2756734187,3733110249),s(3204031479,2999351573),s(3329325298,3815920427),s(3391569614,3928383900),s(3515267271,566280711),s(3940187606,3454069534),s(4118630271,4000239992),s(116418474,1914138554),s(174292421,2731055270),s(289380356,3203993006),s(460393269,320620315),s(685471733,587496836),s(852142971,1086792851),s(1017036298,365543100),s(1126000580,2618297676),s(1288033470,3409855158),s(1501505948,4234509866),s(1607167915,987167468),s(1816402316,1246189591)],a=[];!function(){for(var t=0;t<80;t++)a[t]=s()}();var h=o.SHA512=e.extend({_doReset:function(){this._hash=new n.init([new i.init(1779033703,4089235720),new i.init(3144134277,2227873595),new i.init(1013904242,4271175723),new i.init(2773480762,1595750129),new i.init(1359893119,2917565137),new i.init(2600822924,725511199),new i.init(528734635,4215389547),new i.init(1541459225,327033209)])},_doProcessBlock:function(t,e){for(var r=this._hash.words,i=r[0],n=r[1],o=r[2],s=r[3],h=r[4],l=r[5],f=r[6],d=r[7],u=i.high,p=i.low,_=n.high,v=n.low,y=o.high,g=o.low,B=s.high,w=s.low,k=h.high,S=h.low,m=l.high,x=l.low,b=f.high,H=f.low,z=d.high,A=d.low,C=u,D=p,E=_,R=v,M=y,F=g,P=B,W=w,O=k,I=S,U=m,K=x,X=b,L=H,j=z,N=A,T=0;T<80;T++){var q,Z,V=a[T];if(T<16)Z=V.high=0|t[e+2*T],q=V.low=0|t[e+2*T+1];else{var G=a[T-15],J=G.high,$=G.low,Q=(J>>>1|$<<31)^(J>>>8|$<<24)^J>>>7,Y=($>>>1|J<<31)^($>>>8|J<<24)^($>>>7|J<<25),tt=a[T-2],et=tt.high,rt=tt.low,it=(et>>>19|rt<<13)^(et<<3|rt>>>29)^et>>>6,nt=(rt>>>19|et<<13)^(rt<<3|et>>>29)^(rt>>>6|et<<26),ot=a[T-7],st=ot.high,ct=ot.low,at=a[T-16],ht=at.high,lt=at.low;Z=(Z=(Z=Q+st+((q=Y+ct)>>>0<Y>>>0?1:0))+it+((q+=nt)>>>0<nt>>>0?1:0))+ht+((q+=lt)>>>0<lt>>>0?1:0),V.high=Z,V.low=q}var ft,dt=O&U^~O&X,ut=I&K^~I&L,pt=C&E^C&M^E&M,_t=D&R^D&F^R&F,vt=(C>>>28|D<<4)^(C<<30|D>>>2)^(C<<25|D>>>7),yt=(D>>>28|C<<4)^(D<<30|C>>>2)^(D<<25|C>>>7),gt=(O>>>14|I<<18)^(O>>>18|I<<14)^(O<<23|I>>>9),Bt=(I>>>14|O<<18)^(I>>>18|O<<14)^(I<<23|O>>>9),wt=c[T],kt=wt.high,St=wt.low,mt=j+gt+((ft=N+Bt)>>>0<N>>>0?1:0),xt=yt+_t;j=X,N=L,X=U,L=K,U=O,K=I,O=P+(mt=(mt=(mt=mt+dt+((ft+=ut)>>>0<ut>>>0?1:0))+kt+((ft+=St)>>>0<St>>>0?1:0))+Z+((ft+=q)>>>0<q>>>0?1:0))+((I=W+ft|0)>>>0<W>>>0?1:0)|0,P=M,W=F,M=E,F=R,E=C,R=D,C=mt+(vt+pt+(xt>>>0<yt>>>0?1:0))+((D=ft+xt|0)>>>0<ft>>>0?1:0)|0}p=i.low=p+D,i.high=u+C+(p>>>0<D>>>0?1:0),v=n.low=v+R,n.high=_+E+(v>>>0<R>>>0?1:0),g=o.low=g+F,o.high=y+M+(g>>>0<F>>>0?1:0),w=s.low=w+W,s.high=B+P+(w>>>0<W>>>0?1:0),S=h.low=S+I,h.high=k+O+(S>>>0<I>>>0?1:0),x=l.low=x+K,l.high=m+U+(x>>>0<K>>>0?1:0),H=f.low=H+L,f.high=b+X+(H>>>0<L>>>0?1:0),A=d.low=A+N,d.high=z+j+(A>>>0<N>>>0?1:0)},_doFinalize:function(){var t=this._data,e=t.words,r=8*this._nDataBytes,i=8*t.sigBytes;return e[i>>>5]|=128<<24-i%32,e[30+(128+i>>>10<<5)]=Math.floor(r/4294967296),e[31+(128+i>>>10<<5)]=r,t.sigBytes=4*e.length,this._process(),this._hash.toX32()},clone:function(){var t=e.clone.call(this);return t._hash=this._hash.clone(),t},blockSize:32});t.SHA512=e._createHelper(h),t.HmacSHA512=e._createHmacHelper(h)}(),Z=(q=mt).x64,V=Z.Word,G=Z.WordArray,J=q.algo,$=J.SHA512,Q=J.SHA384=$.extend({_doReset:function(){this._hash=new G.init([new V.init(3418070365,3238371032),new V.init(1654270250,914150663),new V.init(2438529370,812702999),new V.init(355462360,4144912697),new V.init(1731405415,4290775857),new V.init(2394180231,1750603025),new V.init(3675008525,1694076839),new V.init(1203062813,3204075428)])},_doFinalize:function(){var t=$._doFinalize.call(this);return t.sigBytes-=16,t}}),q.SHA384=$._createHelper(Q),q.HmacSHA384=$._createHmacHelper(Q),mt.lib.Cipher||function(){var t=mt,e=t.lib,r=e.Base,i=e.WordArray,n=e.BufferedBlockAlgorithm,o=t.enc,s=(o.Utf8,o.Base64),c=t.algo.EvpKDF,a=e.Cipher=n.extend({cfg:r.extend(),createEncryptor:function(t,e){return this.create(this._ENC_XFORM_MODE,t,e)},createDecryptor:function(t,e){return this.create(this._DEC_XFORM_MODE,t,e)},init:function(t,e,r){this.cfg=this.cfg.extend(r),this._xformMode=t,this._key=e,this.reset()},reset:function(){n.reset.call(this),this._doReset()},process:function(t){return this._append(t),this._process()},finalize:function(t){return t&&this._append(t),this._doFinalize()},keySize:4,ivSize:4,_ENC_XFORM_MODE:1,_DEC_XFORM_MODE:2,_createHelper:function(t){return{encrypt:function(e,r,i){return h(r).encrypt(t,e,r,i)},decrypt:function(e,r,i){return h(r).decrypt(t,e,r,i)}}}});function h(t){return"string"==typeof t?w:g}e.StreamCipher=a.extend({_doFinalize:function(){return this._process(!0)},blockSize:1});var l,f=t.mode={},d=e.BlockCipherMode=r.extend({createEncryptor:function(t,e){return this.Encryptor.create(t,e)},createDecryptor:function(t,e){return this.Decryptor.create(t,e)},init:function(t,e){this._cipher=t,this._iv=e}}),u=f.CBC=((l=d.extend()).Encryptor=l.extend({processBlock:function(t,e){var r=this._cipher,i=r.blockSize;p.call(this,t,e,i),r.encryptBlock(t,e),this._prevBlock=t.slice(e,e+i)}}),l.Decryptor=l.extend({processBlock:function(t,e){var r=this._cipher,i=r.blockSize,n=t.slice(e,e+i);r.decryptBlock(t,e),p.call(this,t,e,i),this._prevBlock=n}}),l);function p(t,e,r){var i,n=this._iv;n?(i=n,this._iv=void 0):i=this._prevBlock;for(var o=0;o<r;o++)t[e+o]^=i[o]}var _=(t.pad={}).Pkcs7={pad:function(t,e){for(var r=4*e,n=r-t.sigBytes%r,o=n<<24|n<<16|n<<8|n,s=[],c=0;c<n;c+=4)s.push(o);var a=i.create(s,n);t.concat(a)},unpad:function(t){var e=255&t.words[t.sigBytes-1>>>2];t.sigBytes-=e}},v=(e.BlockCipher=a.extend({cfg:a.cfg.extend({mode:u,padding:_}),reset:function(){var t;a.reset.call(this);var e=this.cfg,r=e.iv,i=e.mode;this._xformMode==this._ENC_XFORM_MODE?t=i.createEncryptor:(t=i.createDecryptor,this._minBufferSize=1),this._mode&&this._mode.__creator==t?this._mode.init(this,r&&r.words):(this._mode=t.call(i,this,r&&r.words),this._mode.__creator=t)},_doProcessBlock:function(t,e){this._mode.processBlock(t,e)},_doFinalize:function(){var t,e=this.cfg.padding;return this._xformMode==this._ENC_XFORM_MODE?(e.pad(this._data,this.blockSize),t=this._process(!0)):(t=this._process(!0),e.unpad(t)),t},blockSize:4}),e.CipherParams=r.extend({init:function(t){this.mixIn(t)},toString:function(t){return(t||this.formatter).stringify(this)}})),y=(t.format={}).OpenSSL={stringify:function(t){var e=t.ciphertext,r=t.salt;return(r?i.create([1398893684,1701076831]).concat(r).concat(e):e).toString(s)},parse:function(t){var e,r=s.parse(t),n=r.words;return 1398893684==n[0]&&1701076831==n[1]&&(e=i.create(n.slice(2,4)),n.splice(0,4),r.sigBytes-=16),v.create({ciphertext:r,salt:e})}},g=e.SerializableCipher=r.extend({cfg:r.extend({format:y}),encrypt:function(t,e,r,i){i=this.cfg.extend(i);var n=t.createEncryptor(r,i),o=n.finalize(e),s=n.cfg;return v.create({ciphertext:o,key:r,iv:s.iv,algorithm:t,mode:s.mode,padding:s.padding,blockSize:t.blockSize,formatter:i.format})},decrypt:function(t,e,r,i){return i=this.cfg.extend(i),e=this._parse(e,i.format),t.createDecryptor(r,i).finalize(e.ciphertext)},_parse:function(t,e){return"string"==typeof t?e.parse(t,this):t}}),B=(t.kdf={}).OpenSSL={execute:function(t,e,r,n){n=n||i.random(8);var o=c.create({keySize:e+r}).compute(t,n),s=i.create(o.words.slice(e),4*r);return o.sigBytes=4*e,v.create({key:o,iv:s,salt:n})}},w=e.PasswordBasedCipher=g.extend({cfg:g.cfg.extend({kdf:B}),encrypt:function(t,e,r,i){var n=(i=this.cfg.extend(i)).kdf.execute(r,t.keySize,t.ivSize);i.iv=n.iv;var o=g.encrypt.call(this,t,e,n.key,i);return o.mixIn(n),o},decrypt:function(t,e,r,i){i=this.cfg.extend(i),e=this._parse(e,i.format);var n=i.kdf.execute(r,t.keySize,t.ivSize,e.salt);return i.iv=n.iv,g.decrypt.call(this,t,e,n.key,i)}})}(),mt.mode.CFB=((Y=mt.lib.BlockCipherMode.extend()).Encryptor=Y.extend({processBlock:function(t,e){var r=this._cipher,i=r.blockSize;Dt.call(this,t,e,i,r),this._prevBlock=t.slice(e,e+i)}}),Y.Decryptor=Y.extend({processBlock:function(t,e){var r=this._cipher,i=r.blockSize,n=t.slice(e,e+i);Dt.call(this,t,e,i,r),this._prevBlock=n}}),Y),mt.mode.ECB=((tt=mt.lib.BlockCipherMode.extend()).Encryptor=tt.extend({processBlock:function(t,e){this._cipher.encryptBlock(t,e)}}),tt.Decryptor=tt.extend({processBlock:function(t,e){this._cipher.decryptBlock(t,e)}}),tt),mt.pad.AnsiX923={pad:function(t,e){var r=t.sigBytes,i=4*e,n=i-r%i,o=r+n-1;t.clamp(),t.words[o>>>2]|=n<<24-o%4*8,t.sigBytes+=n},unpad:function(t){var e=255&t.words[t.sigBytes-1>>>2];t.sigBytes-=e}},mt.pad.Iso10126={pad:function(t,e){var r=4*e,i=r-t.sigBytes%r;t.concat(mt.lib.WordArray.random(i-1)).concat(mt.lib.WordArray.create([i<<24],1))},unpad:function(t){var e=255&t.words[t.sigBytes-1>>>2];t.sigBytes-=e}},mt.pad.Iso97971={pad:function(t,e){t.concat(mt.lib.WordArray.create([2147483648],1)),mt.pad.ZeroPadding.pad(t,e)},unpad:function(t){mt.pad.ZeroPadding.unpad(t),t.sigBytes--}},mt.mode.OFB=(rt=(et=mt.lib.BlockCipherMode.extend()).Encryptor=et.extend({processBlock:function(t,e){var r=this._cipher,i=r.blockSize,n=this._iv,o=this._keystream;n&&(o=this._keystream=n.slice(0),this._iv=void 0),r.encryptBlock(o,0);for(var s=0;s<i;s++)t[e+s]^=o[s]}}),et.Decryptor=rt,et),mt.pad.NoPadding={pad:function(){},unpad:function(){}},it=mt.lib.CipherParams,nt=mt.enc.Hex,mt.format.Hex={stringify:function(t){return t.ciphertext.toString(nt)},parse:function(t){var e=nt.parse(t);return it.create({ciphertext:e})}},function(){var t=mt,e=t.lib.BlockCipher,r=t.algo,i=[],n=[],o=[],s=[],c=[],a=[],h=[],l=[],f=[],d=[];!function(){for(var t=[],e=0;e<256;e++)t[e]=e<128?e<<1:e<<1^283;var r=0,u=0;for(e=0;e<256;e++){var p=u^u<<1^u<<2^u<<3^u<<4;p=p>>>8^255&p^99,i[r]=p;var _=t[n[p]=r],v=t[_],y=t[v],g=257*t[p]^16843008*p;o[r]=g<<24|g>>>8,s[r]=g<<16|g>>>16,c[r]=g<<8|g>>>24,a[r]=g,g=16843009*y^65537*v^257*_^16843008*r,h[p]=g<<24|g>>>8,l[p]=g<<16|g>>>16,f[p]=g<<8|g>>>24,d[p]=g,r?(r=_^t[t[t[y^_]]],u^=t[t[u]]):r=u=1}}();var u=[0,1,2,4,8,16,32,64,128,27,54],p=r.AES=e.extend({_doReset:function(){if(!this._nRounds||this._keyPriorReset!==this._key){for(var t=this._keyPriorReset=this._key,e=t.words,r=t.sigBytes/4,n=4*(1+(this._nRounds=6+r)),o=this._keySchedule=[],s=0;s<n;s++)s<r?o[s]=e[s]:(p=o[s-1],s%r?6<r&&s%r==4&&(p=i[p>>>24]<<24|i[p>>>16&255]<<16|i[p>>>8&255]<<8|i[255&p]):(p=i[(p=p<<8|p>>>24)>>>24]<<24|i[p>>>16&255]<<16|i[p>>>8&255]<<8|i[255&p],p^=u[s/r|0]<<24),o[s]=o[s-r]^p);for(var c=this._invKeySchedule=[],a=0;a<n;a++){if(s=n-a,a%4)var p=o[s];else p=o[s-4];c[a]=a<4||s<=4?p:h[i[p>>>24]]^l[i[p>>>16&255]]^f[i[p>>>8&255]]^d[i[255&p]]}}},encryptBlock:function(t,e){this._doCryptBlock(t,e,this._keySchedule,o,s,c,a,i)},decryptBlock:function(t,e){var r=t[e+1];t[e+1]=t[e+3],t[e+3]=r,this._doCryptBlock(t,e,this._invKeySchedule,h,l,f,d,n),r=t[e+1],t[e+1]=t[e+3],t[e+3]=r},_doCryptBlock:function(t,e,r,i,n,o,s,c){for(var a=this._nRounds,h=t[e]^r[0],l=t[e+1]^r[1],f=t[e+2]^r[2],d=t[e+3]^r[3],u=4,p=1;p<a;p++){var _=i[h>>>24]^n[l>>>16&255]^o[f>>>8&255]^s[255&d]^r[u++],v=i[l>>>24]^n[f>>>16&255]^o[d>>>8&255]^s[255&h]^r[u++],y=i[f>>>24]^n[d>>>16&255]^o[h>>>8&255]^s[255&l]^r[u++],g=i[d>>>24]^n[h>>>16&255]^o[l>>>8&255]^s[255&f]^r[u++];h=_,l=v,f=y,d=g}_=(c[h>>>24]<<24|c[l>>>16&255]<<16|c[f>>>8&255]<<8|c[255&d])^r[u++],v=(c[l>>>24]<<24|c[f>>>16&255]<<16|c[d>>>8&255]<<8|c[255&h])^r[u++],y=(c[f>>>24]<<24|c[d>>>16&255]<<16|c[h>>>8&255]<<8|c[255&l])^r[u++],g=(c[d>>>24]<<24|c[h>>>16&255]<<16|c[l>>>8&255]<<8|c[255&f])^r[u++],t[e]=_,t[e+1]=v,t[e+2]=y,t[e+3]=g},keySize:8});t.AES=e._createHelper(p)}(),function(){var t=mt,e=t.lib,r=e.WordArray,i=e.BlockCipher,n=t.algo,o=[57,49,41,33,25,17,9,1,58,50,42,34,26,18,10,2,59,51,43,35,27,19,11,3,60,52,44,36,63,55,47,39,31,23,15,7,62,54,46,38,30,22,14,6,61,53,45,37,29,21,13,5,28,20,12,4],s=[14,17,11,24,1,5,3,28,15,6,21,10,23,19,12,4,26,8,16,7,27,20,13,2,41,52,31,37,47,55,30,40,51,45,33,48,44,49,39,56,34,53,46,42,50,36,29,32],c=[1,2,4,6,8,10,12,14,15,17,19,21,23,25,27,28],a=[{0:8421888,268435456:32768,536870912:8421378,805306368:2,1073741824:512,1342177280:8421890,1610612736:8389122,1879048192:8388608,2147483648:514,2415919104:8389120,2684354560:33280,2952790016:8421376,3221225472:32770,3489660928:8388610,3758096384:0,4026531840:33282,134217728:0,402653184:8421890,671088640:33282,939524096:32768,1207959552:8421888,1476395008:512,1744830464:8421378,2013265920:2,2281701376:8389120,2550136832:33280,2818572288:8421376,3087007744:8389122,3355443200:8388610,3623878656:32770,3892314112:514,4160749568:8388608,1:32768,268435457:2,536870913:8421888,805306369:8388608,1073741825:8421378,1342177281:33280,1610612737:512,1879048193:8389122,2147483649:8421890,2415919105:8421376,2684354561:8388610,2952790017:33282,3221225473:514,3489660929:8389120,3758096385:32770,4026531841:0,134217729:8421890,402653185:8421376,671088641:8388608,939524097:512,1207959553:32768,1476395009:8388610,1744830465:2,2013265921:33282,2281701377:32770,2550136833:8389122,2818572289:514,3087007745:8421888,3355443201:8389120,3623878657:0,3892314113:33280,4160749569:8421378},{0:1074282512,16777216:16384,33554432:524288,50331648:1074266128,67108864:1073741840,83886080:1074282496,100663296:1073758208,117440512:16,134217728:540672,150994944:1073758224,167772160:1073741824,184549376:540688,201326592:524304,218103808:0,234881024:16400,251658240:1074266112,8388608:1073758208,25165824:540688,41943040:16,58720256:1073758224,75497472:1074282512,92274688:1073741824,109051904:524288,125829120:1074266128,142606336:524304,159383552:0,176160768:16384,192937984:1074266112,209715200:1073741840,226492416:540672,243269632:1074282496,260046848:16400,268435456:0,285212672:1074266128,301989888:1073758224,318767104:1074282496,335544320:1074266112,352321536:16,369098752:540688,385875968:16384,402653184:16400,419430400:524288,436207616:524304,452984832:1073741840,469762048:540672,486539264:1073758208,503316480:1073741824,520093696:1074282512,276824064:540688,293601280:524288,310378496:1074266112,327155712:16384,343932928:1073758208,360710144:1074282512,377487360:16,394264576:1073741824,411041792:1074282496,427819008:1073741840,444596224:1073758224,461373440:524304,478150656:0,494927872:16400,511705088:1074266128,528482304:540672},{0:260,1048576:0,2097152:67109120,3145728:65796,4194304:65540,5242880:67108868,6291456:67174660,7340032:67174400,8388608:67108864,9437184:67174656,10485760:65792,11534336:67174404,12582912:67109124,13631488:65536,14680064:4,15728640:256,524288:67174656,1572864:67174404,2621440:0,3670016:67109120,4718592:67108868,5767168:65536,6815744:65540,7864320:260,8912896:4,9961472:256,11010048:67174400,12058624:65796,13107200:65792,14155776:67109124,15204352:67174660,16252928:67108864,16777216:67174656,17825792:65540,18874368:65536,19922944:67109120,20971520:256,22020096:67174660,23068672:67108868,24117248:0,25165824:67109124,26214400:67108864,27262976:4,28311552:65792,29360128:67174400,30408704:260,31457280:65796,32505856:67174404,17301504:67108864,18350080:260,19398656:67174656,20447232:0,21495808:65540,22544384:67109120,23592960:256,24641536:67174404,25690112:65536,26738688:67174660,27787264:65796,28835840:67108868,29884416:67109124,30932992:67174400,31981568:4,33030144:65792},{0:2151682048,65536:2147487808,131072:4198464,196608:2151677952,262144:0,327680:4198400,393216:2147483712,458752:4194368,524288:2147483648,589824:4194304,655360:64,720896:2147487744,786432:2151678016,851968:4160,917504:4096,983040:2151682112,32768:2147487808,98304:64,163840:2151678016,229376:2147487744,294912:4198400,360448:2151682112,425984:0,491520:2151677952,557056:4096,622592:2151682048,688128:4194304,753664:4160,819200:2147483648,884736:4194368,950272:4198464,1015808:2147483712,1048576:4194368,1114112:4198400,1179648:2147483712,1245184:0,1310720:4160,1376256:2151678016,1441792:2151682048,1507328:2147487808,1572864:2151682112,1638400:2147483648,1703936:2151677952,1769472:4198464,1835008:2147487744,1900544:4194304,1966080:64,2031616:4096,1081344:2151677952,1146880:2151682112,1212416:0,1277952:4198400,1343488:4194368,1409024:2147483648,1474560:2147487808,1540096:64,1605632:2147483712,1671168:4096,1736704:2147487744,1802240:2151678016,1867776:4160,1933312:2151682048,1998848:4194304,2064384:4198464},{0:128,4096:17039360,8192:262144,12288:536870912,16384:537133184,20480:16777344,24576:553648256,28672:262272,32768:16777216,36864:537133056,40960:536871040,45056:553910400,49152:553910272,53248:0,57344:17039488,61440:553648128,2048:17039488,6144:553648256,10240:128,14336:17039360,18432:262144,22528:537133184,26624:553910272,30720:536870912,34816:537133056,38912:0,43008:553910400,47104:16777344,51200:536871040,55296:553648128,59392:16777216,63488:262272,65536:262144,69632:128,73728:536870912,77824:553648256,81920:16777344,86016:553910272,90112:537133184,94208:16777216,98304:553910400,102400:553648128,106496:17039360,110592:537133056,114688:262272,118784:536871040,122880:0,126976:17039488,67584:553648256,71680:16777216,75776:17039360,79872:537133184,83968:536870912,88064:17039488,92160:128,96256:553910272,100352:262272,104448:553910400,108544:0,112640:553648128,116736:16777344,120832:262144,124928:537133056,129024:536871040},{0:268435464,256:8192,512:270532608,768:270540808,1024:268443648,1280:2097152,1536:2097160,1792:268435456,2048:0,2304:268443656,2560:2105344,2816:8,3072:270532616,3328:2105352,3584:8200,3840:270540800,128:270532608,384:270540808,640:8,896:2097152,1152:2105352,1408:268435464,1664:268443648,1920:8200,2176:2097160,2432:8192,2688:268443656,2944:270532616,3200:0,3456:270540800,3712:2105344,3968:268435456,4096:268443648,4352:270532616,4608:270540808,4864:8200,5120:2097152,5376:268435456,5632:268435464,5888:2105344,6144:2105352,6400:0,6656:8,6912:270532608,7168:8192,7424:268443656,7680:270540800,7936:2097160,4224:8,4480:2105344,4736:2097152,4992:268435464,5248:268443648,5504:8200,5760:270540808,6016:270532608,6272:270540800,6528:270532616,6784:8192,7040:2105352,7296:2097160,7552:0,7808:268435456,8064:268443656},{0:1048576,16:33555457,32:1024,48:1049601,64:34604033,80:0,96:1,112:34603009,128:33555456,144:1048577,160:33554433,176:34604032,192:34603008,208:1025,224:1049600,240:33554432,8:34603009,24:0,40:33555457,56:34604032,72:1048576,88:33554433,104:33554432,120:1025,136:1049601,152:33555456,168:34603008,184:1048577,200:1024,216:34604033,232:1,248:1049600,256:33554432,272:1048576,288:33555457,304:34603009,320:1048577,336:33555456,352:34604032,368:1049601,384:1025,400:34604033,416:1049600,432:1,448:0,464:34603008,480:33554433,496:1024,264:1049600,280:33555457,296:34603009,312:1,328:33554432,344:1048576,360:1025,376:34604032,392:33554433,408:34603008,424:0,440:34604033,456:1049601,472:1024,488:33555456,504:1048577},{0:134219808,1:131072,2:134217728,3:32,4:131104,5:134350880,6:134350848,7:2048,8:134348800,9:134219776,10:133120,11:134348832,12:2080,13:0,14:134217760,15:133152,2147483648:2048,2147483649:134350880,2147483650:134219808,2147483651:134217728,2147483652:134348800,2147483653:133120,2147483654:133152,2147483655:32,2147483656:134217760,2147483657:2080,2147483658:131104,2147483659:134350848,2147483660:0,2147483661:134348832,2147483662:134219776,2147483663:131072,16:133152,17:134350848,18:32,19:2048,20:134219776,21:134217760,22:134348832,23:131072,24:0,25:131104,26:134348800,27:134219808,28:134350880,29:133120,30:2080,31:134217728,2147483664:131072,2147483665:2048,2147483666:134348832,2147483667:133152,2147483668:32,2147483669:134348800,2147483670:134217728,2147483671:134219808,2147483672:134350880,2147483673:134217760,2147483674:134219776,2147483675:0,2147483676:133120,2147483677:2080,2147483678:131104,2147483679:134350848}],h=[4160749569,528482304,33030144,2064384,129024,8064,504,2147483679],l=n.DES=i.extend({_doReset:function(){for(var t=this._key.words,e=[],r=0;r<56;r++){var i=o[r]-1;e[r]=t[i>>>5]>>>31-i%32&1}for(var n=this._subKeys=[],a=0;a<16;a++){var h=n[a]=[],l=c[a];for(r=0;r<24;r++)h[r/6|0]|=e[(s[r]-1+l)%28]<<31-r%6,h[4+(r/6|0)]|=e[28+(s[r+24]-1+l)%28]<<31-r%6;for(h[0]=h[0]<<1|h[0]>>>31,r=1;r<7;r++)h[r]=h[r]>>>4*(r-1)+3;h[7]=h[7]<<5|h[7]>>>27}var f=this._invSubKeys=[];for(r=0;r<16;r++)f[r]=n[15-r]},encryptBlock:function(t,e){this._doCryptBlock(t,e,this._subKeys)},decryptBlock:function(t,e){this._doCryptBlock(t,e,this._invSubKeys)},_doCryptBlock:function(t,e,r){this._lBlock=t[e],this._rBlock=t[e+1],f.call(this,4,252645135),f.call(this,16,65535),d.call(this,2,858993459),d.call(this,8,16711935),f.call(this,1,1431655765);for(var i=0;i<16;i++){for(var n=r[i],o=this._lBlock,s=this._rBlock,c=0,l=0;l<8;l++)c|=a[l][((s^n[l])&h[l])>>>0];this._lBlock=s,this._rBlock=o^c}var u=this._lBlock;this._lBlock=this._rBlock,this._rBlock=u,f.call(this,1,1431655765),d.call(this,8,16711935),d.call(this,2,858993459),f.call(this,16,65535),f.call(this,4,252645135),t[e]=this._lBlock,t[e+1]=this._rBlock},keySize:2,ivSize:2,blockSize:2});function f(t,e){var r=(this._lBlock>>>t^this._rBlock)&e;this._rBlock^=r,this._lBlock^=r<<t}function d(t,e){var r=(this._rBlock>>>t^this._lBlock)&e;this._lBlock^=r,this._rBlock^=r<<t}t.DES=i._createHelper(l);var u=n.TripleDES=i.extend({_doReset:function(){var t=this._key.words;if(2!==t.length&&4!==t.length&&t.length<6)throw new Error("Invalid key length - 3DES requires the key length to be 64, 128, 192 or >192.");var e=t.slice(0,2),i=t.length<4?t.slice(0,2):t.slice(2,4),n=t.length<6?t.slice(0,2):t.slice(4,6);this._des1=l.createEncryptor(r.create(e)),this._des2=l.createEncryptor(r.create(i)),this._des3=l.createEncryptor(r.create(n))},encryptBlock:function(t,e){this._des1.encryptBlock(t,e),this._des2.decryptBlock(t,e),this._des3.encryptBlock(t,e)},decryptBlock:function(t,e){this._des3.decryptBlock(t,e),this._des2.encryptBlock(t,e),this._des1.decryptBlock(t,e)},keySize:6,ivSize:2,blockSize:2});t.TripleDES=i._createHelper(u)}(),function(){var t=mt,e=t.lib.StreamCipher,r=t.algo,i=r.RC4=e.extend({_doReset:function(){for(var t=this._key,e=t.words,r=t.sigBytes,i=this._S=[],n=0;n<256;n++)i[n]=n;n=0;for(var o=0;n<256;n++){var s=n%r,c=e[s>>>2]>>>24-s%4*8&255;o=(o+i[n]+c)%256;var a=i[n];i[n]=i[o],i[o]=a}this._i=this._j=0},_doProcessBlock:function(t,e){t[e]^=n.call(this)},keySize:8,ivSize:0});function n(){for(var t=this._S,e=this._i,r=this._j,i=0,n=0;n<4;n++){r=(r+t[e=(e+1)%256])%256;var o=t[e];t[e]=t[r],t[r]=o,i|=t[(t[e]+t[r])%256]<<24-8*n}return this._i=e,this._j=r,i}t.RC4=e._createHelper(i);var o=r.RC4Drop=i.extend({cfg:i.cfg.extend({drop:192}),_doReset:function(){i._doReset.call(this);for(var t=this.cfg.drop;0<t;t--)n.call(this)}});t.RC4Drop=e._createHelper(o)}(),mt.mode.CTRGladman=(st=(ot=mt.lib.BlockCipherMode.extend()).Encryptor=ot.extend({processBlock:function(t,e){var r,i=this._cipher,n=i.blockSize,o=this._iv,s=this._counter;o&&(s=this._counter=o.slice(0),this._iv=void 0),0===((r=s)[0]=Et(r[0]))&&(r[1]=Et(r[1]));var c=s.slice(0);i.encryptBlock(c,0);for(var a=0;a<n;a++)t[e+a]^=c[a]}}),ot.Decryptor=st,ot),at=(ct=mt).lib.StreamCipher,ht=ct.algo,lt=[],ft=[],dt=[],ut=ht.Rabbit=at.extend({_doReset:function(){for(var t=this._key.words,e=this.cfg.iv,r=0;r<4;r++)t[r]=16711935&(t[r]<<8|t[r]>>>24)|4278255360&(t[r]<<24|t[r]>>>8);var i=this._X=[t[0],t[3]<<16|t[2]>>>16,t[1],t[0]<<16|t[3]>>>16,t[2],t[1]<<16|t[0]>>>16,t[3],t[2]<<16|t[1]>>>16],n=this._C=[t[2]<<16|t[2]>>>16,4294901760&t[0]|65535&t[1],t[3]<<16|t[3]>>>16,4294901760&t[1]|65535&t[2],t[0]<<16|t[0]>>>16,4294901760&t[2]|65535&t[3],t[1]<<16|t[1]>>>16,4294901760&t[3]|65535&t[0]];for(r=this._b=0;r<4;r++)Rt.call(this);for(r=0;r<8;r++)n[r]^=i[r+4&7];if(e){var o=e.words,s=o[0],c=o[1],a=16711935&(s<<8|s>>>24)|4278255360&(s<<24|s>>>8),h=16711935&(c<<8|c>>>24)|4278255360&(c<<24|c>>>8),l=a>>>16|4294901760&h,f=h<<16|65535&a;for(n[0]^=a,n[1]^=l,n[2]^=h,n[3]^=f,n[4]^=a,n[5]^=l,n[6]^=h,n[7]^=f,r=0;r<4;r++)Rt.call(this)}},_doProcessBlock:function(t,e){var r=this._X;Rt.call(this),lt[0]=r[0]^r[5]>>>16^r[3]<<16,lt[1]=r[2]^r[7]>>>16^r[5]<<16,lt[2]=r[4]^r[1]>>>16^r[7]<<16,lt[3]=r[6]^r[3]>>>16^r[1]<<16;for(var i=0;i<4;i++)lt[i]=16711935&(lt[i]<<8|lt[i]>>>24)|4278255360&(lt[i]<<24|lt[i]>>>8),t[e+i]^=lt[i]},blockSize:4,ivSize:2}),ct.Rabbit=at._createHelper(ut),mt.mode.CTR=(_t=(pt=mt.lib.BlockCipherMode.extend()).Encryptor=pt.extend({processBlock:function(t,e){var r=this._cipher,i=r.blockSize,n=this._iv,o=this._counter;n&&(o=this._counter=n.slice(0),this._iv=void 0);var s=o.slice(0);r.encryptBlock(s,0),o[i-1]=o[i-1]+1|0;for(var c=0;c<i;c++)t[e+c]^=s[c]}}),pt.Decryptor=_t,pt),yt=(vt=mt).lib.StreamCipher,gt=vt.algo,Bt=[],wt=[],kt=[],St=gt.RabbitLegacy=yt.extend({_doReset:function(){for(var t=this._key.words,e=this.cfg.iv,r=this._X=[t[0],t[3]<<16|t[2]>>>16,t[1],t[0]<<16|t[3]>>>16,t[2],t[1]<<16|t[0]>>>16,t[3],t[2]<<16|t[1]>>>16],i=this._C=[t[2]<<16|t[2]>>>16,4294901760&t[0]|65535&t[1],t[3]<<16|t[3]>>>16,4294901760&t[1]|65535&t[2],t[0]<<16|t[0]>>>16,4294901760&t[2]|65535&t[3],t[1]<<16|t[1]>>>16,4294901760&t[3]|65535&t[0]],n=this._b=0;n<4;n++)Mt.call(this);for(n=0;n<8;n++)i[n]^=r[n+4&7];if(e){var o=e.words,s=o[0],c=o[1],a=16711935&(s<<8|s>>>24)|4278255360&(s<<24|s>>>8),h=16711935&(c<<8|c>>>24)|4278255360&(c<<24|c>>>8),l=a>>>16|4294901760&h,f=h<<16|65535&a;for(i[0]^=a,i[1]^=l,i[2]^=h,i[3]^=f,i[4]^=a,i[5]^=l,i[6]^=h,i[7]^=f,n=0;n<4;n++)Mt.call(this)}},_doProcessBlock:function(t,e){var r=this._X;Mt.call(this),Bt[0]=r[0]^r[5]>>>16^r[3]<<16,Bt[1]=r[2]^r[7]>>>16^r[5]<<16,Bt[2]=r[4]^r[1]>>>16^r[7]<<16,Bt[3]=r[6]^r[3]>>>16^r[1]<<16;for(var i=0;i<4;i++)Bt[i]=16711935&(Bt[i]<<8|Bt[i]>>>24)|4278255360&(Bt[i]<<24|Bt[i]>>>8),t[e+i]^=Bt[i]},blockSize:4,ivSize:2}),vt.RabbitLegacy=yt._createHelper(St),mt.pad.ZeroPadding={pad:function(t,e){var r=4*e;t.clamp(),t.sigBytes+=r-(t.sigBytes%r||r)},unpad:function(t){var e=t.words,r=t.sigBytes-1;for(r=t.sigBytes-1;0<=r;r--)if(e[r>>>2]>>>24-r%4*8&255){t.sigBytes=r+1;break}}},mt});
 }
-
 
 // prettier-ignore
 function Env(t,e){"undefined"!=typeof process&&JSON.stringify(process.env).indexOf("GITHUB")>-1&&process.exit(0);class s{constructor(t){this.env=t}send(t,e="GET"){t="string"==typeof t?{url:t}:t;let s=this.get;return"POST"===e&&(s=this.post),new Promise((e,i)=>{s.call(this,t,(t,s,r)=>{t?i(t):e(s)})})}get(t){return this.send.call(this.env,t)}post(t){return this.send.call(this.env,t,"POST")}}return new class{constructor(t,e){this.name=t,this.http=new s(this),this.data=null,this.dataFile="box.dat",this.logs=[],this.isMute=!1,this.isNeedRewrite=!1,this.logSeparator="\n",this.startTime=(new Date).getTime(),Object.assign(this,e),this.log("",`🔔${this.name}, 开始!`)}isNode(){return"undefined"!=typeof module&&!!module.exports}isQuanX(){return"undefined"!=typeof $task}isSurge(){return"undefined"!=typeof $httpClient&&"undefined"==typeof $loon}isLoon(){return"undefined"!=typeof $loon}toObj(t,e=null){try{return JSON.parse(t)}catch{return e}}toStr(t,e=null){try{return JSON.stringify(t)}catch{return e}}getjson(t,e){let s=e;const i=this.getdata(t);if(i)try{s=JSON.parse(this.getdata(t))}catch{}return s}setjson(t,e){try{return this.setdata(JSON.stringify(t),e)}catch{return!1}}getScript(t){return new Promise(e=>{this.get({url:t},(t,s,i)=>e(i))})}runScript(t,e){return new Promise(s=>{let i=this.getdata("@chavy_boxjs_userCfgs.httpapi");i=i?i.replace(/\n/g,"").trim():i;let r=this.getdata("@chavy_boxjs_userCfgs.httpapi_timeout");r=r?1*r:20,r=e&&e.timeout?e.timeout:r;const[o,h]=i.split("@"),n={url:`http://${h}/v1/scripting/evaluate`,body:{script_text:t,mock_type:"cron",timeout:r},headers:{"X-Key":o,Accept:"*/*"}};this.post(n,(t,e,i)=>s(i))}).catch(t=>this.logErr(t))}loaddata(){if(!this.isNode())return{};{this.fs=this.fs?this.fs:require("fs"),this.path=this.path?this.path:require("path");const t=this.path.resolve(this.dataFile),e=this.path.resolve(process.cwd(),this.dataFile),s=this.fs.existsSync(t),i=!s&&this.fs.existsSync(e);if(!s&&!i)return{};{const i=s?t:e;try{return JSON.parse(this.fs.readFileSync(i))}catch(t){return{}}}}}writedata(){if(this.isNode()){this.fs=this.fs?this.fs:require("fs"),this.path=this.path?this.path:require("path");const t=this.path.resolve(this.dataFile),e=this.path.resolve(process.cwd(),this.dataFile),s=this.fs.existsSync(t),i=!s&&this.fs.existsSync(e),r=JSON.stringify(this.data);s?this.fs.writeFileSync(t,r):i?this.fs.writeFileSync(e,r):this.fs.writeFileSync(t,r)}}lodash_get(t,e,s){const i=e.replace(/\[(\d+)\]/g,".$1").split(".");let r=t;for(const t of i)if(r=Object(r)[t],void 0===r)return s;return r}lodash_set(t,e,s){return Object(t)!==t?t:(Array.isArray(e)||(e=e.toString().match(/[^.[\]]+/g)||[]),e.slice(0,-1).reduce((t,s,i)=>Object(t[s])===t[s]?t[s]:t[s]=Math.abs(e[i+1])>>0==+e[i+1]?[]:{},t)[e[e.length-1]]=s,t)}getdata(t){let e=this.getval(t);if(/^@/.test(t)){const[,s,i]=/^@(.*?)\.(.*?)$/.exec(t),r=s?this.getval(s):"";if(r)try{const t=JSON.parse(r);e=t?this.lodash_get(t,i,""):e}catch(t){e=""}}return e}setdata(t,e){let s=!1;if(/^@/.test(e)){const[,i,r]=/^@(.*?)\.(.*?)$/.exec(e),o=this.getval(i),h=i?"null"===o?null:o||"{}":"{}";try{const e=JSON.parse(h);this.lodash_set(e,r,t),s=this.setval(JSON.stringify(e),i)}catch(e){const o={};this.lodash_set(o,r,t),s=this.setval(JSON.stringify(o),i)}}else s=this.setval(t,e);return s}getval(t){return this.isSurge()||this.isLoon()?$persistentStore.read(t):this.isQuanX()?$prefs.valueForKey(t):this.isNode()?(this.data=this.loaddata(),this.data[t]):this.data&&this.data[t]||null}setval(t,e){return this.isSurge()||this.isLoon()?$persistentStore.write(t,e):this.isQuanX()?$prefs.setValueForKey(t,e):this.isNode()?(this.data=this.loaddata(),this.data[e]=t,this.writedata(),!0):this.data&&this.data[e]||null}initGotEnv(t){this.got=this.got?this.got:require("got"),this.cktough=this.cktough?this.cktough:require("tough-cookie"),this.ckjar=this.ckjar?this.ckjar:new this.cktough.CookieJar,t&&(t.headers=t.headers?t.headers:{},void 0===t.headers.Cookie&&void 0===t.cookieJar&&(t.cookieJar=this.ckjar))}get(t,e=(()=>{})){t.headers&&(delete t.headers["Content-Type"],delete t.headers["Content-Length"]),this.isSurge()||this.isLoon()?(this.isSurge()&&this.isNeedRewrite&&(t.headers=t.headers||{},Object.assign(t.headers,{"X-Surge-Skip-Scripting":!1})),$httpClient.get(t,(t,s,i)=>{!t&&s&&(s.body=i,s.statusCode=s.status),e(t,s,i)})):this.isQuanX()?(this.isNeedRewrite&&(t.opts=t.opts||{},Object.assign(t.opts,{hints:!1})),$task.fetch(t).then(t=>{const{statusCode:s,statusCode:i,headers:r,body:o}=t;e(null,{status:s,statusCode:i,headers:r,body:o},o)},t=>e(t))):this.isNode()&&(this.initGotEnv(t),this.got(t).on("redirect",(t,e)=>{try{if(t.headers["set-cookie"]){const s=t.headers["set-cookie"].map(this.cktough.Cookie.parse).toString();s&&this.ckjar.setCookieSync(s,null),e.cookieJar=this.ckjar}}catch(t){this.logErr(t)}}).then(t=>{const{statusCode:s,statusCode:i,headers:r,body:o}=t;e(null,{status:s,statusCode:i,headers:r,body:o},o)},t=>{const{message:s,response:i}=t;e(s,i,i&&i.body)}))}post(t,e=(()=>{})){if(t.body&&t.headers&&!t.headers["Content-Type"]&&(t.headers["Content-Type"]="application/x-www-form-urlencoded"),t.headers&&delete t.headers["Content-Length"],this.isSurge()||this.isLoon())this.isSurge()&&this.isNeedRewrite&&(t.headers=t.headers||{},Object.assign(t.headers,{"X-Surge-Skip-Scripting":!1})),$httpClient.post(t,(t,s,i)=>{!t&&s&&(s.body=i,s.statusCode=s.status),e(t,s,i)});else if(this.isQuanX())t.method="POST",this.isNeedRewrite&&(t.opts=t.opts||{},Object.assign(t.opts,{hints:!1})),$task.fetch(t).then(t=>{const{statusCode:s,statusCode:i,headers:r,body:o}=t;e(null,{status:s,statusCode:i,headers:r,body:o},o)},t=>e(t));else if(this.isNode()){this.initGotEnv(t);const{url:s,...i}=t;this.got.post(s,i).then(t=>{const{statusCode:s,statusCode:i,headers:r,body:o}=t;e(null,{status:s,statusCode:i,headers:r,body:o},o)},t=>{const{message:s,response:i}=t;e(s,i,i&&i.body)})}}time(t,e=null){const s=e?new Date(e):new Date;let i={"M+":s.getMonth()+1,"d+":s.getDate(),"H+":s.getHours(),"m+":s.getMinutes(),"s+":s.getSeconds(),"q+":Math.floor((s.getMonth()+3)/3),S:s.getMilliseconds()};/(y+)/.test(t)&&(t=t.replace(RegExp.$1,(s.getFullYear()+"").substr(4-RegExp.$1.length)));for(let e in i)new RegExp("("+e+")").test(t)&&(t=t.replace(RegExp.$1,1==RegExp.$1.length?i[e]:("00"+i[e]).substr((""+i[e]).length)));return t}msg(e=t,s="",i="",r){const o=t=>{if(!t)return t;if("string"==typeof t)return this.isLoon()?t:this.isQuanX()?{"open-url":t}:this.isSurge()?{url:t}:void 0;if("object"==typeof t){if(this.isLoon()){let e=t.openUrl||t.url||t["open-url"],s=t.mediaUrl||t["media-url"];return{openUrl:e,mediaUrl:s}}if(this.isQuanX()){let e=t["open-url"]||t.url||t.openUrl,s=t["media-url"]||t.mediaUrl;return{"open-url":e,"media-url":s}}if(this.isSurge()){let e=t.url||t.openUrl||t["open-url"];return{url:e}}}};if(this.isMute||(this.isSurge()||this.isLoon()?$notification.post(e,s,i,o(r)):this.isQuanX()&&$notify(e,s,i,o(r))),!this.isMuteLog){let t=["","==============📣系统通知📣=============="];t.push(e),s&&t.push(s),i&&t.push(i),console.log(t.join("\n")),this.logs=this.logs.concat(t)}}log(...t){t.length>0&&(this.logs=[...this.logs,...t]),console.log(t.join(this.logSeparator))}logErr(t,e){const s=!this.isSurge()&&!this.isQuanX()&&!this.isLoon();s?this.log("",`❗️${this.name}, 错误!`,t.stack):this.log("",`❗️${this.name}, 错误!`,t)}wait(t){return new Promise(e=>setTimeout(e,t))}done(t={}){const e=(new Date).getTime(),s=(e-this.startTime)/1e3;this.log("",`🔔${this.name}, 结束! 🕛 ${s} 秒`),this.log(),(this.isSurge()||this.isQuanX()||this.isLoon())&&$done(t)}}(t,e)}
